% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{data}
\name{dm_curso}
\alias{dm_curso}
\title{Dados do censo ensino superior INEP 2017 (Cursos)}
\description{
Base de dados do INEP para o censo brasileiro do ensino superior (cursos).
}
\examples{
require(rnp)
str(dm_curso)
}
\references{
\url{http://inep.gov.br/microdados}
}
\author{
INEP
}
\keyword{dm_curso}
