% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\docType{data}
\name{dm_local}
\alias{dm_local}
\title{Dados do censo ensino superior INEP 2017 (Local da oferta)}
\description{
Base de dados do INEP para o censo brasileiro do ensino superior (locais de oferta).
}
\examples{
require(rnp)
str(dm_local)
}
\references{
\url{http://inep.gov.br/microdados}
}
\author{
INEP
}
\keyword{dm_local}
