% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{media_aritmetica}
\alias{media_aritmetica}
\title{Media aritmetica simples e ponderada}
\usage{
media_aritmetica(x, peso = NULL, remove.na = TRUE)
}
\arguments{
\item{x}{vetor numerico.}

\item{peso}{vetor de pesos. Se informado, calcula a media ponderada  .}

\item{remove.na}{Se TRUE remove NA's do vetor x e/ou dos pesos.}
}
\value{
numerico
}
\description{
Media aritmetica simples e ponderada
}
\details{
Esta funcao e um atalho para as funcoes \code{\link{mean}} e \code{\link{weighted.mean}}
que calculam respectivamente as medias simples e ponderada. Caso seja passado um vetor de pesos e
ele contenha missing e o argumento remove.na for TRUE constuimos um data.frame e removemos os
NA's na mesma posicao que aparecem nos dois vetores.
}
\examples{
require(rnp)
x <- c(2,2,5)
p <- c(3,5,NA)
media_aritmetica(x)
media_aritmetica(x, p)
media_aritmetica(x, p, remove.na = TRUE)
}
\author{
LOPES, J. E.
}
