% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{media_geometrica}
\alias{media_geometrica}
\title{Media geometrica simples e ponderada}
\usage{
media_geometrica(x, peso = NULL, remove.na = TRUE)
}
\arguments{
\item{x}{vetor numerico.}

\item{peso}{vetor de pesos. Se informado, calcula a media ponderada  .}

\item{remove.na}{Se TRUE remove NA's do vetor x e/ou dos pesos.}
}
\value{
numerico
}
\description{
Media geometrica simples e ponderada
}
\details{
Esta funcao calcula respectivamente as medias geometricas simples e ponderada.
Caso seja passado um vetor de pesos e ele contenha missing e o argumento remove.na
for TRUE constuimos um data.frame e removemos os NA's na mesma posicao que aparecem
nos dois vetores.
}
\examples{
require(rnp)
x <- c(2,2,5)
p <- c(3,5,NA)
media_geometrica(x)
media_geometrica(x, p)
media_geometrica(x, p, remove.na = TRUE)
}
\author{
LOPES, J. E.
}
