% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{media_harmonica}
\alias{media_harmonica}
\title{Media harmonica simples e ponderada}
\usage{
media_harmonica(x, peso = NULL, remove.na = TRUE)
}
\arguments{
\item{x}{vetor numerico.}

\item{peso}{vetor de pesos. Se informado, calcula a media ponderada  .}

\item{remove.na}{Se TRUE remove NA's do vetor x e/ou dos pesos.}
}
\value{
numerico
}
\description{
Media harmonica simples e ponderada
}
\details{
Esta funcao calcula respectivamente as medias harmonicas simples e ponderada.
Caso seja passado um vetor de pesos e ele contenha missing e o argumento remove.na
for TRUE constuimos um data.frame e removemos os NA's na mesma posicao que aparecem
nos dois vetores.
}
\examples{
require(rnp)
x <- c(2,2,5)
p <- c(3,5,NA)
media_harmonica(x)
media_harmonica(x, p)
media_harmonica(x, p, remove.na = TRUE)
}
\author{
LOPES, J. E.
}
