% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp-package.R
\docType{package}
\name{rnp-package}
\alias{rnp}
\alias{rnp-package}
\title{rnp: Recursos estatisticos para o projeto R NA PRATICA}
\description{
Pacote didatico com funcoes estatisticas de producao: descritiva,
inferencia, regressao, multivariada, series temporais, probabilidade,
amostragem e visualizacao ggplot2. Inclui ferramentas para dados do
censo do ensino superior do INEP.
}
\section{Modulos}{

\itemize{
\item \code{\link[=rnp_descritiva]{rnp_descritiva()}}, \code{\link[=rnp_quantis]{rnp_quantis()}}, \code{\link[=rnp_outliers]{rnp_outliers()}}
\item \code{\link[=rnp_tabela_classes]{rnp_tabela_classes()}}, \code{\link[=rnp_distribuicao]{rnp_distribuicao()}}
\item \code{\link[=rnp_ic_media]{rnp_ic_media()}}, \code{\link[=rnp_teste_t]{rnp_teste_t()}}, \code{\link[=rnp_anova]{rnp_anova()}}
\item \code{\link[=rnp_correlacao_teste]{rnp_correlacao_teste()}}, \code{\link[=rnp_regressao]{rnp_regressao()}}
\item \code{\link[=rnp_pca]{rnp_pca()}}, \code{\link[=rnp_kmeans]{rnp_kmeans()}}
\item \code{\link[=rnp_ts_decomposicao]{rnp_ts_decomposicao()}}, \code{\link[=rnp_ts_arima]{rnp_ts_arima()}}
\item \code{\link[=rnp_combinacao]{rnp_combinacao()}}, \code{\link[=rnp_distribuicao_binomial]{rnp_distribuicao_binomial()}}
\item \code{\link[=rnp_grafico_dispersao]{rnp_grafico_dispersao()}}, \code{\link[=rnp_tema_rnp]{rnp_tema_rnp()}}
\item \code{\link[=rnp_amostra_simples]{rnp_amostra_simples()}}, \code{\link[=rnp_tamanho_amostra_media]{rnp_tamanho_amostra_media()}}
\item \code{\link[=rnp_tamanho_efeito]{rnp_tamanho_efeito()}}, \code{\link[=rnp_na_summary]{rnp_na_summary()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evandeilton/rnp}
  \item Report bugs at \url{https://github.com/evandeilton/rnp/issues}
}

}
\author{
Jose E. Lopes \email{evandeilton@gmail.com}
}
\keyword{internal}
