% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliacao.R
\name{rnp_acuracia_diagnostica}
\alias{rnp_acuracia_diagnostica}
\title{Acuracia diagnostica}
\usage{
rnp_acuracia_diagnostica(observado, predito, positivo = NULL, digits = 4L)
}
\arguments{
\item{observado}{Condicao verdadeira (binaria).}

\item{predito}{Resultado do teste (binario).}

\item{positivo}{Nivel "doente"/positivo.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{metrica} e \code{valor}.
}
\description{
Metricas de avaliacao de testes diagnosticos: sensibilidade, especificidade,
valores preditivos e razoes de verossimilhanca.
}
\examples{
obs <- c("D","D","D","S","S","S","S","D")
tst <- c("+","+","-","-","-","+","-","+")
rnp_acuracia_diagnostica(obs, tst, positivo = "D")
}
\seealso{
Other avaliacao: 
\code{\link{rnp_brier}()},
\code{\link{rnp_calibracao}()},
\code{\link{rnp_comparar_roc}()},
\code{\link{rnp_curva_ganho}()},
\code{\link{rnp_curva_lift}()},
\code{\link{rnp_curva_precisao_revocacao}()},
\code{\link{rnp_ks_classificador}()},
\code{\link{rnp_metricas_classificacao}()},
\code{\link{rnp_metricas_regressao}()}
}
\concept{avaliacao}
