% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amostragem.R
\name{rnp_amostra_simples}
\alias{rnp_amostra_simples}
\title{Amostragem aleatoria simples}
\usage{
rnp_amostra_simples(base, n, seed = NULL)
}
\arguments{
\item{base}{data.frame ou vetor.}

\item{n}{Inteiro. Tamanho da amostra.}

\item{seed}{Inteiro. Semente aleatoria.}
}
\value{
Amostra (mesmo tipo de base).
}
\description{
Amostragem aleatoria simples
}
\examples{
rnp_amostra_simples(mtcars, 10)
}
