% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amostragem.R
\name{rnp_amostra_sistematica}
\alias{rnp_amostra_sistematica}
\title{Amostragem sistematica}
\usage{
rnp_amostra_sistematica(base, n, seed = NULL)
}
\arguments{
\item{base}{data.frame ou vetor.}

\item{n}{Inteiro. Tamanho da amostra.}

\item{seed}{Inteiro. Semente aleatoria (para o inicio).}
}
\value{
Amostra.
}
\description{
Amostragem sistematica
}
\examples{
rnp_amostra_sistematica(mtcars, 10)
}
