% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{rnp_ancova}
\alias{rnp_ancova}
\title{Analise de covariancia (ANCOVA)}
\usage{
rnp_ancova(formula, data, digits = 4L)
}
\arguments{
\item{formula}{Formula \code{y ~ fator + covariavel}.}

\item{data}{data.frame.}

\item{digits}{Inteiro.}
}
\value{
tibble com a tabela ANOVA do modelo.
}
\description{
Ajusta um modelo com um fator e uma covariavel continua, reportando a tabela
ANCOVA (tipo I).
}
\examples{
rnp_ancova(mpg ~ factor(cyl) + wt, mtcars)
}
\seealso{
Other experimental: 
\code{\link{rnp_anova_medidas_repetidas}()},
\code{\link{rnp_contrastes}()},
\code{\link{rnp_dbc}()}
}
\concept{experimental}
