% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_anova_dois_fatores}
\alias{rnp_anova_dois_fatores}
\title{ANOVA de dois fatores}
\usage{
rnp_anova_dois_fatores(formula, data = NULL, digits = 4L)
}
\arguments{
\item{formula}{Formula \code{y ~ A * B} ou \code{y ~ A + B}.}

\item{data}{data.frame.}

\item{digits}{Inteiro.}
}
\value{
tibble com a tabela ANOVA expandida.
}
\description{
ANOVA de dois fatores
}
\examples{
rnp_anova_dois_fatores(breaks ~ wool * tension, warpbreaks)
}
