% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao-avancada.R
\name{rnp_anova_modelos}
\alias{rnp_anova_modelos}
\title{Comparacao de modelos aninhados (ANOVA)}
\usage{
rnp_anova_modelos(..., digits = 4L)
}
\arguments{
\item{...}{Modelos \code{lm}/\code{glm} na ordem crescente de complexidade.}

\item{digits}{Inteiro.}
}
\value{
tibble com a tabela de comparacao.
}
\description{
Compara dois ou mais modelos \code{lm}/\code{glm} aninhados via \code{\link[stats:anova]{stats::anova()}}.
}
\examples{
m1 <- lm(mpg ~ wt, mtcars); m2 <- lm(mpg ~ wt + hp, mtcars)
rnp_anova_modelos(m1, m2)
}
\seealso{
Other regressao: 
\code{\link{rnp_box_cox}()},
\code{\link{rnp_elastic_net}()},
\code{\link{rnp_grafico_residuos}()},
\code{\link{rnp_predicao}()},
\code{\link{rnp_regressao_lasso}()},
\code{\link{rnp_regressao_multinomial}()},
\code{\link{rnp_regressao_nao_linear}()},
\code{\link{rnp_regressao_poisson}()},
\code{\link{rnp_regressao_polinomial}()},
\code{\link{rnp_regressao_ponderada}()},
\code{\link{rnp_regressao_ridge}()},
\code{\link{rnp_regressao_robusta}()},
\code{\link{rnp_regressao_stepwise}()},
\code{\link{rnp_vif}()}
}
\concept{regressao}
