% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inep.R
\name{rnp_aplica_classes}
\alias{rnp_aplica_classes}
\title{Aplica classes do dicionario INEP na base}
\usage{
rnp_aplica_classes(base, classes)
}
\arguments{
\item{base}{Base de dados INEP.}

\item{classes}{Base de classes (de \code{\link[=rnp_get_classes_inep]{rnp_get_classes_inep()}}).}
}
\value{
data.frame com colunas _DESC adicionadas.
}
\description{
Junta descricoes das categorias com os codigos numericos.
}
\examples{
\dontrun{
classes <- rnp_get_classes_inep("Dicionario.xlsx", aba = "DM_IES")
base <- rnp_read("DM_IES.csv", sep = "|")
base_desc <- rnp_aplica_classes(base, classes)
}
}
