% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_arranjo}
\alias{rnp_arranjo}
\title{Numero de arranjos}
\usage{
rnp_arranjo(n, k)
}
\arguments{
\item{n}{Inteiro. Total de elementos.}

\item{k}{Inteiro. Tamanho do subconjunto.}
}
\value{
Escalar numerico.
}
\description{
Numero de arranjos
}
\examples{
rnp_arranjo(10, 3)
}
