% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_atributos}
\alias{rnp_atributos}
\title{Extrai atributos de um objeto}
\usage{
rnp_atributos(obj)
}
\arguments{
\item{obj}{qualquer objeto em R}
}
\value{
data.frame com lista de atributos
}
\description{
Extrai atributos de um objeto
}
\author{
LOPES, J. E
}
