% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_auto_arima}
\alias{rnp_auto_arima}
\title{Selecao automatica de ordem ARIMA}
\usage{
rnp_auto_arima(
  x,
  max_p = 3,
  max_d = 2,
  max_q = 3,
  periodo = 1,
  ic = c("aicc", "aic", "bic"),
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico ou \code{ts}.}

\item{max_p, max_d, max_q}{Ordens maximas nao-sazonais.}

\item{periodo}{Periodo sazonal (1 = sem sazonalidade).}

\item{ic}{String: \code{"aicc"}, \code{"aic"} ou \code{"bic"}.}

\item{digits}{Inteiro.}
}
\value{
Uma lista como \code{\link[=rnp_arima]{rnp_arima()}}, com \code{selecao} (tibble dos melhores
candidatos) adicional.
}
\description{
Busca em grade a melhor ordem ARIMA(p, d, q) (opcionalmente sazonal) pelo
criterio escolhido. Substitui \code{forecast::auto.arima} com implementacao
propria sobre \code{\link[stats:arima]{stats::arima()}}.
}
\examples{
rnp_auto_arima(lh, max_p = 2, max_q = 2)$modelo
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
