% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verossimilhanca.R
\name{rnp_bayes_conjugada}
\alias{rnp_bayes_conjugada}
\title{Estimacao bayesiana conjugada}
\usage{
rnp_bayes_conjugada(
  familia = c("beta_binomial", "gama_poisson", "normal_normal"),
  priori,
  dados,
  conf = 0.95,
  digits = 4L
)
}
\arguments{
\item{familia}{String: \code{"beta_binomial"}, \code{"gama_poisson"} ou
\code{"normal_normal"}.}

\item{priori}{Vetor nomeado de hiperparametros a priori
(\code{beta_binomial}: \code{a}, \code{b}; \code{gama_poisson}: \code{a}, \code{b};
\code{normal_normal}: \code{mu0}, \code{sigma0}).}

\item{dados}{Lista com os dados
(\code{beta_binomial}: \code{sucessos}, \code{n}; \code{gama_poisson}: \code{soma}, \code{n};
\code{normal_normal}: \code{x} e \code{sigma} conhecido).}

\item{conf}{Nivel de credibilidade do intervalo.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{parametro}, \code{valor} (hiperparametros a posteriori),
alem de \code{media_post}, \code{ic_inf}, \code{ic_sup}.
}
\description{
Atualiza distribuicoes a priori conjugadas com os dados para tres modelos
classicos: Beta-Binomial, Gama-Poisson e Normal-Normal (variancia conhecida).
}
\examples{
rnp_bayes_conjugada("beta_binomial", priori = c(a = 1, b = 1),
                    dados = list(sucessos = 8, n = 10))
}
\seealso{
Other inferencia: 
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
