% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada-fatorial.R
\name{rnp_biplot}
\alias{rnp_biplot}
\title{Biplot de PCA}
\usage{
rnp_biplot(rnp_pca_obj, escala = 1)
}
\arguments{
\item{rnp_pca_obj}{Saida de \code{\link[=rnp_pca]{rnp_pca()}}.}

\item{escala}{Escala dos vetores de carga (fator multiplicativo).}
}
\value{
Objeto \code{ggplot}.
}
\description{
Constroi um biplot (scores das observacoes + vetores das variaveis) a partir
da saida de \code{\link[=rnp_pca]{rnp_pca()}}.
}
\examples{
p <- rnp_pca(mtcars[, c("mpg", "disp", "hp", "drat", "wt")])
rnp_biplot(p)
}
\seealso{
Other multivariada: 
\code{\link{rnp_analise_fatorial}()},
\code{\link{rnp_cluster_hierarquico}()},
\code{\link{rnp_correlacao_canonica}()},
\code{\link{rnp_correspondencia}()},
\code{\link{rnp_grafico_correlograma}()},
\code{\link{rnp_grafico_dendrograma}()},
\code{\link{rnp_grafico_dispersao_matriz}()},
\code{\link{rnp_hotelling}()},
\code{\link{rnp_kmedoids}()},
\code{\link{rnp_lda}()},
\code{\link{rnp_manova}()},
\code{\link{rnp_matriz_correlacao}()},
\code{\link{rnp_normalidade_multivariada}()},
\code{\link{rnp_silhueta}()}
}
\concept{multivariada}
