% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rnp.R
\docType{data}
\name{rnp_concreto}
\alias{rnp_concreto}
\title{Resistencia de corpos de prova de concreto (didatico)}
\format{
Um \code{data.frame} com 90 linhas e 3 variaveis:
\describe{
\item{resistencia}{Resistencia a compressao (MPa).}
\item{dias_cura}{Tempo de cura (7, 14 ou 28 dias).}
\item{tipo_cimento}{Tipo de cimento (\code{CP-II}, \code{CP-IV}, \code{CP-V}).}
}
}
\usage{
rnp_concreto
}
\description{
Dados \strong{simulados} de ensaios de compressao de corpos de prova de concreto,
para ilustrar estatistica descritiva, ANOVA e regressao.
}
\details{
Dados simulados para fins didaticos; ver \code{data-raw/datasets.R}.
}
\examples{
rnp_descritiva(rnp_concreto$resistencia)
rnp_anova(resistencia ~ tipo_cimento, data = rnp_concreto)
}
\keyword{datasets}
