% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada-avancada.R
\name{rnp_correlacao_canonica}
\alias{rnp_correlacao_canonica}
\title{Correlacao canonica}
\usage{
rnp_correlacao_canonica(X, Y, digits = 4L)
}
\arguments{
\item{X, Y}{Matrizes/data.frames numericos com o mesmo numero de linhas.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{dimensao} e \code{correlacao_canonica}.
}
\description{
Encontra as correlacoes canonicas entre dois conjuntos de variaveis.
}
\examples{
rnp_correlacao_canonica(iris[, 1:2], iris[, 3:4])
}
\seealso{
Other multivariada: 
\code{\link{rnp_analise_fatorial}()},
\code{\link{rnp_biplot}()},
\code{\link{rnp_cluster_hierarquico}()},
\code{\link{rnp_correspondencia}()},
\code{\link{rnp_grafico_correlograma}()},
\code{\link{rnp_grafico_dendrograma}()},
\code{\link{rnp_grafico_dispersao_matriz}()},
\code{\link{rnp_hotelling}()},
\code{\link{rnp_kmedoids}()},
\code{\link{rnp_lda}()},
\code{\link{rnp_manova}()},
\code{\link{rnp_matriz_correlacao}()},
\code{\link{rnp_normalidade_multivariada}()},
\code{\link{rnp_silhueta}()}
}
\concept{multivariada}
