% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{rnp_correlacao_parcial}
\alias{rnp_correlacao_parcial}
\title{Correlacao parcial}
\usage{
rnp_correlacao_parcial(x, y, z, method = "pearson", digits = 4L)
}
\arguments{
\item{x, y}{Vetores numericos.}

\item{z}{Vetor numerico ou matriz de variaveis de controle.}

\item{method}{String. Apenas \code{"pearson"} implementado por enquanto.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{correlacao_parcial}, \code{gl}, \code{p_valor}, \code{n}.
}
\description{
Calcula correlacao parcial entre x e y, controlando por z (vetor ou matriz).
}
\examples{
rnp_correlacao_parcial(mtcars$mpg, mtcars$hp, mtcars$wt)
}
