% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobrevivencia.R
\name{rnp_cox}
\alias{rnp_cox}
\title{Modelo de Cox de riscos proporcionais}
\usage{
rnp_cox(formula, data, conf = 0.95, digits = 4L)
}
\arguments{
\item{formula}{Formula, ex.: \code{Surv(time, status) ~ age + sex}.}

\item{data}{data.frame.}

\item{conf}{Nivel de confianca.}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{coeficientes} (tibble com \code{hazard_ratio}, IC, p),
\code{modelo} (concordancia, AIC, testes globais) e \code{objeto} (\code{coxph}).
}
\description{
Ajusta o modelo semiparametrico de Cox. A formula deve ter
\code{survival::Surv(tempo, evento)} no lado esquerdo.
}
\examples{
rnp_cox(survival::Surv(time, status) ~ age + sex + ph.ecog,
        data = survival::lung)$coeficientes
}
\seealso{
Other sobrevivencia: 
\code{\link{rnp_cox_diagnosticos}()},
\code{\link{rnp_cox_risco_relativo}()},
\code{\link{rnp_grafico_sobrevivencia}()},
\code{\link{rnp_kaplan_meier}()},
\code{\link{rnp_log_rank}()},
\code{\link{rnp_nelson_aalen}()},
\code{\link{rnp_sobrevivencia_parametrica}()},
\code{\link{rnp_tabela_vida}()}
}
\concept{sobrevivencia}
