% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobrevivencia.R
\name{rnp_cox_risco_relativo}
\alias{rnp_cox_risco_relativo}
\title{Risco relativo predito pelo modelo de Cox}
\usage{
rnp_cox_risco_relativo(modelo, novos_dados, digits = 4L)
}
\arguments{
\item{modelo}{Objeto \code{coxph} ou saida de \code{\link[=rnp_cox]{rnp_cox()}}.}

\item{novos_dados}{data.frame com as covariaveis dos perfis.}

\item{digits}{Inteiro.}
}
\value{
tibble com o risco relativo predito por perfil.
}
\description{
Calcula o risco relativo (em relacao a media das covariaveis) para novos
perfis de covariaveis.
}
\examples{
fit <- rnp_cox(survival::Surv(time, status) ~ age + sex, survival::lung)
rnp_cox_risco_relativo(fit, data.frame(age = c(50, 70), sex = c(1, 1)))
}
\seealso{
Other sobrevivencia: 
\code{\link{rnp_cox}()},
\code{\link{rnp_cox_diagnosticos}()},
\code{\link{rnp_grafico_sobrevivencia}()},
\code{\link{rnp_kaplan_meier}()},
\code{\link{rnp_log_rank}()},
\code{\link{rnp_nelson_aalen}()},
\code{\link{rnp_sobrevivencia_parametrica}()},
\code{\link{rnp_tabela_vida}()}
}
\concept{sobrevivencia}
