% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessamento.R
\name{rnp_discretiza}
\alias{rnp_discretiza}
\title{Discretizacao (binning)}
\usage{
rnp_discretiza(
  x,
  k = 4L,
  metodo = c("largura", "frequencia", "kmeans"),
  rotulos = NULL
)
}
\arguments{
\item{x}{Vetor numerico.}

\item{k}{Inteiro. Numero de classes.}

\item{metodo}{String: \code{"largura"}, \code{"frequencia"} ou \code{"kmeans"}.}

\item{rotulos}{Vetor opcional de rotulos das classes.}
}
\value{
Fator com \code{k} niveis.
}
\description{
Converte uma variavel continua em categorias por largura igual, frequencia
igual (quantis) ou k-means.
}
\examples{
rnp_discretiza(mtcars$mpg, k = 3, metodo = "frequencia")
}
\seealso{
Other preprocessamento: 
\code{\link{rnp_dummy}()},
\code{\link{rnp_imputa}()},
\code{\link{rnp_normaliza}()},
\code{\link{rnp_padroniza}()},
\code{\link{rnp_winsoriza}()}
}
\concept{preprocessamento}
