% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada.R
\name{rnp_distancia}
\alias{rnp_distancia}
\title{Distancias entre observacoes}
\usage{
rnp_distancia(
  base,
  method = c("euclidean", "manhattan", "minkowski", "canberra", "mahalanobis"),
  p = 2,
  digits = 4L
)
}
\arguments{
\item{base}{data.frame ou matriz numerica.}

\item{method}{String: \code{"euclidean"}, \code{"manhattan"},
\code{"minkowski"}, \code{"canberra"}, \code{"mahalanobis"}.}

\item{p}{Escalar. Potencia (apenas Minkowski).}

\item{digits}{Inteiro.}
}
\value{
Matriz de distancias (classe \code{dist}).
}
\description{
Calcula matriz de distancias usando varios metodos.
}
\examples{
rnp_distancia(mtcars[1:5, c("mpg", "hp")], method = "euclidean")
}
