% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_beta}
\alias{rnp_distribuicao_beta}
\title{Distribuicao Beta}
\usage{
rnp_distribuicao_beta(
  fun = c("d", "p", "q", "r"),
  a,
  b,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{a, b}{Parametros de forma (shape1, shape2), positivos.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Beta
}
\examples{
rnp_distribuicao_beta("d", x = 0.5, a = 2, b = 2)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
