% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_gama}
\alias{rnp_distribuicao_gama}
\title{Distribuicao Gama}
\usage{
rnp_distribuicao_gama(
  fun = c("d", "p", "q", "r"),
  forma,
  taxa = 1,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{forma, taxa}{Parametros de forma (shape) e taxa (rate), positivos.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Gama
}
\examples{
rnp_distribuicao_gama("d", x = 2, forma = 2, taxa = 1)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
