% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_distribuicao_geom}
\alias{rnp_distribuicao_geom}
\title{Distribuicao geometrica (wrapper)}
\usage{
rnp_distribuicao_geom(
  fun = c("d", "p", "q", "r"),
  prob,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"} (massa), \code{"p"} (cumulativa),
\code{"q"} (quantil), \code{"r"} (amostra).}

\item{prob}{Probabilidade de sucesso em cada ensaio.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao geometrica (wrapper)
}
\examples{
rnp_distribuicao_geom("d", x = 0:5, prob = .2)
}
