% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_lognormal}
\alias{rnp_distribuicao_lognormal}
\title{Distribuicao Log-normal}
\usage{
rnp_distribuicao_lognormal(
  fun = c("d", "p", "q", "r"),
  media_log = 0,
  dp_log = 1,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{media_log, dp_log}{Media e desvio-padrao na escala logaritmica.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Log-normal
}
\examples{
rnp_distribuicao_lognormal("d", x = 1)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
