% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_multinomial}
\alias{rnp_distribuicao_multinomial}
\title{Distribuicao Multinomial}
\usage{
rnp_distribuicao_multinomial(
  fun = c("d", "r"),
  tamanho,
  probs,
  x = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"} (massa conjunta) ou \code{"r"} (amostra).}

\item{tamanho}{Numero de ensaios (size).}

\item{probs}{Vetor de probabilidades das categorias (soma 1).}

\item{x}{Vetor de contagens (para \code{fun = "d"}), de mesmo tamanho de \code{probs}.}

\item{n}{Numero de amostras (para \code{fun = "r"}).}
}
\value{
Escalar (d) ou matriz de contagens (r).
}
\description{
Distribuicao Multinomial
}
\examples{
rnp_distribuicao_multinomial("d", tamanho = 10, probs = c(.2, .3, .5),
                             x = c(2, 3, 5))
rnp_distribuicao_multinomial("r", tamanho = 10, probs = c(.2, .3, .5), n = 4)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
