% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_distribuicao_negbinom}
\alias{rnp_distribuicao_negbinom}
\title{Distribuicao binomial negativa (wrapper)}
\usage{
rnp_distribuicao_negbinom(
  fun = c("d", "p", "q", "r"),
  size,
  prob,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"} (massa), \code{"p"} (cumulativa),
\code{"q"} (quantil), \code{"r"} (amostra).}

\item{size}{Numero de sucessos alvo.}

\item{prob}{Probabilidade de sucesso em cada ensaio.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao binomial negativa (wrapper)
}
\examples{
rnp_distribuicao_negbinom("d", x = 0:5, size = 3, prob = .5)
}
