% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_normal}
\alias{rnp_distribuicao_normal}
\title{Distribuicao Normal}
\usage{
rnp_distribuicao_normal(
  fun = c("d", "p", "q", "r"),
  media = 0,
  dp = 1,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{media, dp}{Media e desvio-padrao.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Normal
}
\examples{
rnp_distribuicao_normal("d", x = 0)
rnp_distribuicao_normal("p", q = 1.96)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
