% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_distribuicao_poisson}
\alias{rnp_distribuicao_poisson}
\title{Distribuicao de Poisson (wrapper)}
\usage{
rnp_distribuicao_poisson(
  fun = c("d", "p", "q", "r"),
  lambda,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"} (massa), \code{"p"} (cumulativa),
\code{"q"} (quantil), \code{"r"} (amostra).}

\item{lambda}{Taxa media (positivo).}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao de Poisson (wrapper)
}
\examples{
rnp_distribuicao_poisson("d", x = 0:5, lambda = 2)
}
