% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_qui_quadrado}
\alias{rnp_distribuicao_qui_quadrado}
\title{Distribuicao Qui-quadrado}
\usage{
rnp_distribuicao_qui_quadrado(
  fun = c("d", "p", "q", "r"),
  gl,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{gl}{Graus de liberdade, positivo.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Qui-quadrado
}
\examples{
rnp_distribuicao_qui_quadrado("q", p = 0.95, gl = 3)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_distribuicao_weibull}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
