% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuicoes-continuas.R
\name{rnp_distribuicao_weibull}
\alias{rnp_distribuicao_weibull}
\title{Distribuicao Weibull}
\usage{
rnp_distribuicao_weibull(
  fun = c("d", "p", "q", "r"),
  forma,
  escala = 1,
  x = NULL,
  q = NULL,
  p = NULL,
  n = NULL
)
}
\arguments{
\item{fun}{String: \code{"d"}, \code{"p"}, \code{"q"}, \code{"r"}.}

\item{forma, escala}{Parametros de forma (shape) e escala (scale), positivos.}

\item{x, q, p, n}{Argumentos conforme \code{fun}.}
}
\value{
Vetor numerico.
}
\description{
Distribuicao Weibull
}
\examples{
rnp_distribuicao_weibull("p", q = 1, forma = 1.5, escala = 1)
}
\seealso{
Other distribuicoes: 
\code{\link{rnp_ajuste_distribuicao}()},
\code{\link{rnp_distribuicao_beta}()},
\code{\link{rnp_distribuicao_exponencial}()},
\code{\link{rnp_distribuicao_f}()},
\code{\link{rnp_distribuicao_gama}()},
\code{\link{rnp_distribuicao_lognormal}()},
\code{\link{rnp_distribuicao_multinomial}()},
\code{\link{rnp_distribuicao_normal}()},
\code{\link{rnp_distribuicao_qui_quadrado}()},
\code{\link{rnp_distribuicao_t}()},
\code{\link{rnp_distribuicao_uniforme}()},
\code{\link{rnp_grafico_distribuicao}()}
}
\concept{distribuicoes}
