% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verossimilhanca.R
\name{rnp_emv}
\alias{rnp_emv}
\title{Estimador de maxima verossimilhanca (generico)}
\usage{
rnp_emv(log_veross, inicio, nomes = NULL, metodo = "Nelder-Mead", digits = 4L)
}
\arguments{
\item{log_veross}{Funcao \code{function(theta) ...} que devolve a
log-verossimilhanca (escalar) para o vetor de parametros \code{theta}.}

\item{inicio}{Vetor de valores iniciais para \code{theta}.}

\item{nomes}{Vetor opcional de nomes dos parametros.}

\item{metodo}{Metodo de \code{\link[stats:optim]{stats::optim()}}.}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{estimativas} (tibble: \code{parametro}, \code{estimativa},
\code{erro_padrao}, \code{z}, \code{ic_inf}, \code{ic_sup}) e \code{ajuste} (tibble:
\code{log_veross}, \code{aic}, \code{bic}... \code{convergiu}).
}
\description{
Maximiza uma funcao de log-verossimilhanca via \code{\link[stats:optim]{stats::optim()}} e estima
erros-padrao pela inversa da informacao de Fisher observada (Hessiana).
}
\examples{
set.seed(1); x <- rnorm(200, 5, 2)
ll <- function(th) sum(dnorm(x, th[1], th[2], log = TRUE))
rnp_emv(ll, inicio = c(0, 1), nomes = c("media", "dp"))$estimativas
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
