% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_estrutura}
\alias{rnp_estrutura}
\title{Estrutura de um objeto (glance)}
\usage{
rnp_estrutura(obj)
}
\arguments{
\item{obj}{Vetor, \code{data.frame} ou \code{tibble}.}
}
\value{
tibble com \code{variavel}, \code{classe}, \code{n}, \code{n_faltantes},
\code{p_faltantes}.
}
\description{
Resume a estrutura de um vetor, \code{data.frame} ou \code{tibble}: classe de cada
variavel, numero de observacoes e contagem de valores faltantes. Substitui
a antiga \code{rnp_atributos}, agora com saida em \code{tibble}.
}
\examples{
rnp_estrutura(mtcars)
rnp_estrutura(airquality$Ozone)
}
