% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inep.R
\name{rnp_get_classes_inep}
\alias{rnp_get_classes_inep}
\title{Obtem classes das variaveis do dicionario INEP}
\usage{
rnp_get_classes_inep(caminho, aba = 1, pula_linha = 1, retorna_lista = FALSE)
}
\arguments{
\item{caminho}{Caminho do arquivo Excel.}

\item{aba}{Nome ou numero da aba.}

\item{pula_linha}{Linhas a pular.}

\item{retorna_lista}{Logico.}
}
\value{
data.frame.
}
\description{
Extrai categorias e descricoes do dicionario de dados do INEP.
}
\examples{
\dontrun{
rnp_get_classes_inep("Dicionario_de_Dados.xlsx", aba = "DM_IES")
}
}
