% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inep.R
\name{rnp_get_inep_censo}
\alias{rnp_get_inep_censo}
\title{Download dados do INEP}
\usage{
rnp_get_inep_censo(ano = 2020, url = NULL, salvar = NULL)
}
\arguments{
\item{ano}{Ano (1995-2023).}

\item{url}{URL customizada (sobrescreve ano).}

\item{salvar}{Caminho de destino.}
}
\value{
Arquivo .zip baixado.
}
\description{
Baixa microdados do censo do ensino superior do INEP.
}
\examples{
\dontrun{
rnp_get_inep_censo(ano = 2020)
}
}
