% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_grafico_acf}
\alias{rnp_grafico_acf}
\title{Grafico de ACF ou PACF}
\usage{
rnp_grafico_acf(x, tipo = c("acf", "pacf"), lag_max = NULL)
}
\arguments{
\item{x}{Vetor numerico ou \code{ts}.}

\item{tipo}{String: \code{"acf"} ou \code{"pacf"}.}

\item{lag_max}{Inteiro. Numero maximo de defasagens.}
}
\value{
Objeto \code{ggplot}.
}
\description{
Grafico de ACF ou PACF
}
\examples{
rnp_grafico_acf(as.numeric(AirPassengers), tipo = "acf")
}
\seealso{
Other series: 
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()}
}
\concept{series}
