% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_grafico_barras}
\alias{rnp_grafico_barras}
\title{Grafico de barras}
\usage{
rnp_grafico_barras(
  base,
  x,
  y = NULL,
  fill = NULL,
  position = c("dodge", "stack", "fill"),
  titulo = NULL,
  xlab = NULL,
  ylab = NULL,
  tema = rnp_tema_rnp()
)
}
\arguments{
\item{base}{data.frame.}

\item{x}{Nome da coluna (string) para eixo x.}

\item{y}{Nome opcional da coluna (string) para eixo y. Se NULL, conta.}

\item{fill}{Nome opcional da coluna para preenchimento.}

\item{position}{String: \code{"dodge"}, \code{"stack"}, \code{"fill"}.}

\item{titulo}{String. Titulo do grafico.}

\item{xlab, ylab}{String. Rotulos dos eixos.}

\item{tema}{Objeto \code{theme} do ggplot2. Default \code{rnp_tema_rnp()}.}
}
\value{
Objeto ggplot.
}
\description{
Wrapper ggplot2 para grafico de barras.
}
\examples{
rnp_grafico_barras(mtcars, "cyl")
}
