% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada-avancada.R
\name{rnp_grafico_dendrograma}
\alias{rnp_grafico_dendrograma}
\title{Dendrograma}
\usage{
rnp_grafico_dendrograma(cluster_hier, k = NULL)
}
\arguments{
\item{cluster_hier}{Saida de \code{\link[=rnp_cluster_hierarquico]{rnp_cluster_hierarquico()}} ou objeto \code{hclust}.}

\item{k}{Inteiro opcional. Desenha \code{k} retangulos de corte (informativo).}
}
\value{
Objeto \code{ggplot}.
}
\description{
Dendrograma
}
\examples{
ch <- rnp_cluster_hierarquico(mtcars[, c("mpg", "hp", "wt")], k = 3)
rnp_grafico_dendrograma(ch)
}
\seealso{
Other multivariada: 
\code{\link{rnp_analise_fatorial}()},
\code{\link{rnp_biplot}()},
\code{\link{rnp_cluster_hierarquico}()},
\code{\link{rnp_correlacao_canonica}()},
\code{\link{rnp_correspondencia}()},
\code{\link{rnp_grafico_correlograma}()},
\code{\link{rnp_grafico_dispersao_matriz}()},
\code{\link{rnp_hotelling}()},
\code{\link{rnp_kmedoids}()},
\code{\link{rnp_lda}()},
\code{\link{rnp_manova}()},
\code{\link{rnp_matriz_correlacao}()},
\code{\link{rnp_normalidade_multivariada}()},
\code{\link{rnp_silhueta}()}
}
\concept{multivariada}
