% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_grafico_dispersao}
\alias{rnp_grafico_dispersao}
\title{Grafico de dispersao}
\usage{
rnp_grafico_dispersao(
  base,
  x,
  y,
  cor = NULL,
  tamanho = 2,
  suavizar = FALSE,
  titulo = NULL,
  xlab = NULL,
  ylab = NULL,
  tema = rnp_tema_rnp()
)
}
\arguments{
\item{base}{data.frame.}

\item{x, y}{Nomes das colunas (strings).}

\item{cor}{Nome opcional da coluna para cor.}

\item{tamanho}{Escalar ou nome de coluna para tamanho dos pontos.}

\item{suavizar}{Logico. Adicionar linha de suavizacao (loess).}

\item{titulo, xlab, ylab}{String.}

\item{tema}{Objeto \code{theme}.}
}
\value{
Objeto ggplot.
}
\description{
Grafico de dispersao
}
\examples{
rnp_grafico_dispersao(mtcars, "wt", "mpg")
}
