% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_grafico_histograma}
\alias{rnp_grafico_histograma}
\title{Grafico de histograma}
\usage{
rnp_grafico_histograma(
  base,
  x,
  bins = 30,
  fill = rnp_paleta_rnp("rnp_qual", 1),
  densidade = FALSE,
  titulo = NULL,
  xlab = NULL,
  ylab = NULL,
  tema = rnp_tema_rnp()
)
}
\arguments{
\item{base}{data.frame.}

\item{x}{Nome da coluna (string).}

\item{bins}{Inteiro. Numero de bins.}

\item{fill}{Cor de preenchimento.}

\item{densidade}{Logico. Mostrar densidade em vez de contagem.}

\item{titulo, xlab, ylab}{String.}

\item{tema}{Objeto \code{theme}.}
}
\value{
Objeto ggplot.
}
\description{
Grafico de histograma
}
\examples{
rnp_grafico_histograma(mtcars, "mpg")
}
