% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_grafico_qq}
\alias{rnp_grafico_qq}
\title{Grafico QQ (quantil-quantil)}
\usage{
rnp_grafico_qq(
  x,
  distribuicao = "norm",
  ...,
  titulo = NULL,
  xlab = "Quantis Teoricos",
  ylab = "Quantis Amostrais",
  tema = rnp_tema_rnp()
)
}
\arguments{
\item{x}{Vetor numerico.}

\item{distribuicao}{String: \code{"norm"}, \code{"t"}, \code{"chisq"}, etc.}

\item{...}{Argumentos da distribuicao (ex.: \code{df} para t).}

\item{titulo, xlab, ylab}{String.}

\item{tema}{Objeto \code{theme}.}
}
\value{
Objeto ggplot.
}
\description{
Grafico QQ (quantil-quantil)
}
\examples{
rnp_grafico_qq(rnorm(100))
}
