% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao-avancada.R
\name{rnp_grafico_residuos}
\alias{rnp_grafico_residuos}
\title{Painel de diagnostico de residuos}
\usage{
rnp_grafico_residuos(modelo)
}
\arguments{
\item{modelo}{Objeto \code{lm}.}
}
\value{
Uma lista nomeada de objetos \code{ggplot}: \code{residuo_ajustado}, \code{qq},
\code{escala_locacao}, \code{residuo_leverage}.
}
\description{
Gera os graficos classicos de diagnostico de um \code{lm}.
}
\examples{
g <- rnp_grafico_residuos(lm(mpg ~ wt + hp, mtcars))
g$residuo_ajustado
}
\seealso{
Other regressao: 
\code{\link{rnp_anova_modelos}()},
\code{\link{rnp_box_cox}()},
\code{\link{rnp_elastic_net}()},
\code{\link{rnp_predicao}()},
\code{\link{rnp_regressao_lasso}()},
\code{\link{rnp_regressao_multinomial}()},
\code{\link{rnp_regressao_nao_linear}()},
\code{\link{rnp_regressao_poisson}()},
\code{\link{rnp_regressao_polinomial}()},
\code{\link{rnp_regressao_ponderada}()},
\code{\link{rnp_regressao_ridge}()},
\code{\link{rnp_regressao_robusta}()},
\code{\link{rnp_regressao_stepwise}()},
\code{\link{rnp_vif}()}
}
\concept{regressao}
