% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_grafico_serie}
\alias{rnp_grafico_serie}
\title{Grafico de serie temporal}
\usage{
rnp_grafico_serie(x, titulo = NULL)
}
\arguments{
\item{x}{Vetor numerico ou \code{ts}.}

\item{titulo}{Titulo opcional.}
}
\value{
Objeto \code{ggplot}.
}
\description{
Grafico de serie temporal
}
\examples{
rnp_grafico_serie(as.numeric(AirPassengers))
}
\seealso{
Other series: 
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()}
}
\concept{series}
