% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobrevivencia.R
\name{rnp_grafico_sobrevivencia}
\alias{rnp_grafico_sobrevivencia}
\title{Curva de Kaplan-Meier (grafico)}
\usage{
rnp_grafico_sobrevivencia(km, intervalo = TRUE, titulo = NULL)
}
\arguments{
\item{km}{Saida de \code{\link[=rnp_kaplan_meier]{rnp_kaplan_meier()}} (ou objeto \code{survfit}).}

\item{intervalo}{Logico. Exibe a faixa de confianca.}

\item{titulo}{Titulo opcional.}
}
\value{
Objeto \code{ggplot}.
}
\description{
Curva de Kaplan-Meier (grafico)
}
\examples{
km <- rnp_kaplan_meier(survival::lung$time, survival::lung$status == 2,
                       grupo = survival::lung$sex)
rnp_grafico_sobrevivencia(km)
}
\seealso{
Other sobrevivencia: 
\code{\link{rnp_cox}()},
\code{\link{rnp_cox_diagnosticos}()},
\code{\link{rnp_cox_risco_relativo}()},
\code{\link{rnp_kaplan_meier}()},
\code{\link{rnp_log_rank}()},
\code{\link{rnp_nelson_aalen}()},
\code{\link{rnp_sobrevivencia_parametrica}()},
\code{\link{rnp_tabela_vida}()}
}
\concept{sobrevivencia}
