% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_grafico_violino}
\alias{rnp_grafico_violino}
\title{Grafico de violino}
\usage{
rnp_grafico_violino(
  base,
  x,
  y,
  fill = NULL,
  titulo = NULL,
  xlab = NULL,
  ylab = NULL,
  tema = rnp_tema_rnp()
)
}
\arguments{
\item{base}{data.frame.}

\item{x}{Nome da coluna (string) para eixo x (categorica).}

\item{y}{Nome da coluna (string) para eixo y (numerica).}

\item{fill}{Nome opcional da coluna para preenchimento.}

\item{titulo, xlab, ylab}{String.}

\item{tema}{Objeto \code{theme}.}
}
\value{
Objeto ggplot.
}
\description{
Grafico de violino
}
\examples{
rnp_grafico_violino(mtcars, "cyl", "mpg")
}
