% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencia.R
\name{rnp_ic_diff_medias}
\alias{rnp_ic_diff_medias}
\title{Intervalo de confianca para diferenca de medias}
\usage{
rnp_ic_diff_medias(
  x,
  y,
  pareado = FALSE,
  var_iguais = FALSE,
  conf = 0.95,
  na.rm = TRUE,
  digits = 4L
)
}
\arguments{
\item{x, y}{Vetores numericos.}

\item{pareado}{Logico. Amostras pareadas.}

\item{var_iguais}{Logico. Assumir variancias iguais (independentes).}

\item{conf}{Nivel de confianca.}

\item{na.rm}{Logico.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{diff_medias}, \code{erro_padrao},
\code{limite_inferior}, \code{limite_superior}, \code{gl},
\code{metodo}.
}
\description{
Intervalo de confianca para diferenca de medias
}
\examples{
rnp_ic_diff_medias(rnorm(30, 5), rnorm(30, 5.5))
}
