% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verossimilhanca.R
\name{rnp_informacao_fisher}
\alias{rnp_informacao_fisher}
\title{Informacao de Fisher observada}
\usage{
rnp_informacao_fisher(log_veross, theta, digits = 4L)
}
\arguments{
\item{log_veross}{Funcao \code{function(theta) ...} (log-verossimilhanca).}

\item{theta}{Ponto de avaliacao (tipicamente a EMV).}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{informacao} (matriz) e \code{erros_padrao} (vetor).
}
\description{
Calcula a matriz de informacao de Fisher observada (Hessiana negativa da
log-verossimilhanca) em um ponto, e os erros-padrao assintoticos.
}
\examples{
set.seed(1); x <- rnorm(100, 2, 1)
ll <- function(th) sum(dnorm(x, th[1], th[2], log = TRUE))
rnp_informacao_fisher(ll, c(2, 1))$erros_padrao
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
