% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_kruskal}
\alias{rnp_kruskal}
\title{Teste de Kruskal-Wallis}
\usage{
rnp_kruskal(x, g = NULL, digits = 4L, data = NULL)
}
\arguments{
\item{x}{Vetor numerico ou formula.}

\item{g}{Vetor de grupos.}

\item{digits}{Inteiro.}

\item{data}{data.frame opcional (usado quando \code{x} e uma formula).}
}
\value{
tibble.
}
\description{
Teste de Kruskal-Wallis
}
\examples{
rnp_kruskal(mtcars$mpg, as.factor(mtcars$cyl))
}
