% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_load_packages}
\alias{rnp_load_packages}
\title{Carregar ou instalar pacotes CRAN}
\usage{
rnp_load_packages(pacotes = c("tidyverse", "lubridate", "magrittr",
  "broom", "stringr", "plotly", "ggplot2", "data.table", "DT", "formatR",
  "svglite", "tufte", "kableExtra", "ggmosaic", "prettydoc"))
}
\arguments{
\item{pacotes}{Lista ou vetr e nomes de pacotes separados por aspas e ponto e virgula.}
}
\description{
Carega pacotes necesários pra trabalhar com o R NA PRATICA
}
\examples{
require(rnp)
rnp_load_packages()
}
