% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verossimilhanca.R
\name{rnp_log_verossimilhanca}
\alias{rnp_log_verossimilhanca}
\title{Grafico da funcao de log-verossimilhanca (1 parametro)}
\usage{
rnp_log_verossimilhanca(log_veross, intervalo, n_pontos = 200L, titulo = NULL)
}
\arguments{
\item{log_veross}{Funcao escalar \code{function(theta) ...}.}

\item{intervalo}{Vetor \code{c(min, max)} do parametro.}

\item{n_pontos}{Inteiro. Resolucao da curva.}

\item{titulo}{Titulo opcional.}
}
\value{
Objeto \code{ggplot}.
}
\description{
Plota a log-verossimilhanca em funcao de um parametro escalar num intervalo,
marcando a estimativa de maxima verossimilhanca.
}
\examples{
set.seed(1); x <- rpois(100, 3)
rnp_log_verossimilhanca(function(l) sum(dpois(x, l, log = TRUE)),
                        intervalo = c(1, 6))
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
