% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_mann_whitney}
\alias{rnp_mann_whitney}
\title{Teste de Mann-Whitney (Wilcoxon duas amostras independentes)}
\usage{
rnp_mann_whitney(
  x,
  y,
  lado = c("bilateral", "direita", "esquerda"),
  digits = 4L
)
}
\arguments{
\item{x, y}{Vetores numericos.}

\item{lado}{String.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Teste de Mann-Whitney (Wilcoxon duas amostras independentes)
}
\examples{
rnp_mann_whitney(rnorm(20, 5), rnorm(20, 6))
}
