% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada-avancada.R
\name{rnp_manova}
\alias{rnp_manova}
\title{MANOVA (analise de variancia multivariada)}
\usage{
rnp_manova(formula, data, digits = 4L)
}
\arguments{
\item{formula}{Formula \code{cbind(y1, y2, ...) ~ grupo}.}

\item{data}{data.frame.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{teste}, \code{estatistica}, \code{aprox_f}, \code{p_valor}.
}
\description{
Wrapper de \code{\link[stats:manova]{stats::manova()}} reportando as estatisticas de Wilks e Pillai.
}
\examples{
rnp_manova(cbind(Sepal.Length, Petal.Length) ~ Species, iris)
}
\seealso{
Other multivariada: 
\code{\link{rnp_analise_fatorial}()},
\code{\link{rnp_biplot}()},
\code{\link{rnp_cluster_hierarquico}()},
\code{\link{rnp_correlacao_canonica}()},
\code{\link{rnp_correspondencia}()},
\code{\link{rnp_grafico_correlograma}()},
\code{\link{rnp_grafico_dendrograma}()},
\code{\link{rnp_grafico_dispersao_matriz}()},
\code{\link{rnp_hotelling}()},
\code{\link{rnp_kmedoids}()},
\code{\link{rnp_lda}()},
\code{\link{rnp_matriz_correlacao}()},
\code{\link{rnp_normalidade_multivariada}()},
\code{\link{rnp_silhueta}()}
}
\concept{multivariada}
