% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada.R
\name{rnp_mds}
\alias{rnp_mds}
\title{Escalonamento multidimensional (MDS)}
\usage{
rnp_mds(d, k = 2, digits = 4L)
}
\arguments{
\item{d}{Objeto \code{dist} ou matriz de distancias.}

\item{k}{Inteiro. Numero de dimensoes.}

\item{digits}{Inteiro.}
}
\value{
lista:
\itemize{
\item \code{pontos}: tibble com coordenadas.
\item \code{eigenvalues}: vetor.
\item \code{GOF}: vetor com goodness-of-fit.
}
}
\description{
Wrapper de \code{\link[stats:cmdscale]{stats::cmdscale()}}.
}
\examples{
d <- dist(mtcars[1:10, c("mpg", "hp", "wt")])
rnp_mds(d, k = 2)
}
