% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_media}
\alias{rnp_media}
\title{Calcula medias de varios tipos}
\usage{
rnp_media(x, peso = NULL, remove.na = TRUE)
}
\arguments{
\item{x}{Vetor, matrix ou data.frame de dados numericos}

\item{peso}{Opcional. Vetor, matrix ou data.frame de dados numericos com pesos com as mesmas
caracteristcias de comprimento ou dimensoes do argumento x}

\item{remove.na}{Se TRUE remove os NA's, caso existam nos dados}
}
\value{
Vetor ou data.frame com as medias de uma ou mais variaveis. No caso de data.frames, a
saida e um data.frame com quatro colunas, sendo uma o nome das variaveis e as outras tres as
colunas de medias aritmetica, geometrica e harmonica respectivamente. No caso de vetor numerico,
a saida e um data.frame simples com tres colunas e uma linha contendo as medias.
}
\description{
Esta funcao e uma chamada master paras as funcoes \code{\link{media_aritmetica}},
\code{\link{media_geometrica}} e \code{\link{media_harmonica}}. Serve para calcular médias
de vetores numéricos e de matrizes/data.frames formados por colunas numericas.
}
\details{
Atraves desta funcao e possivel calcular medias aritmeticas, geometricas e harmonicas
simples e ponderadas. Se for passado uma matriz ela varre tdoas as colunas e determina as medias
para cada variavel. Se for passado argumento peso, as medias sao ponderadas pelos seus respectivos
pesos. No caso de matrizes, a matriz de peso precisa ter as mesmas dimensoes da matriz de origem
e o mesmo tipo de dados.
Quando existir peso, o nome das colunas do data.frame serao nomeadas como: media_ap, media_gp e media_hp.
Caso contrario, media_a, media_g e media_h.
}
\examples{
require(rnp)
# Exemplos com vetores
x <- mtcars$mpg;x[3:5] <- NA
p <- mtcars$cyl;p[5:7] <- NA
rnp_media(x = x, peso = NULL, remove.na = TRUE)
rnp_media(x = x, peso = NULL, remove.na = FALSE)
rnp_media(x = x, peso = p, remove.na = TRUE)
rnp_media(x = x, peso = p, remove.na = FALSE)
# Exemplos com matrizes (de mesmo tipo)
X <- mtcars
Y <- data.frame(matrix(runif(prod(dim(mtcars)), 5, 10), nrow = nrow(mtcars), ncol = ncol(mtcars)));
class(X); class(Y)
names(Y) <- names(X)
dim(X) == dim(Y)
rnp_media(X)
rnp_media(X, Y)
# Com dados reais da base rnp::dm_ies
X <- as.data.frame(rnp::dm_ies)
X <- X[,c(sapply(X, class) \%in\% c("integer", "numeric"))]
rnp_media(X[,-c(1:6)])
a <- sapply(X[,-c(1:6)], mean)
b <- rnp_media(X[,-c(1:6)])$media_a
all(a == b)
}
\author{
LOPES, J. E.
}
