% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliacao.R
\name{rnp_metricas_regressao}
\alias{rnp_metricas_regressao}
\title{Metricas de regressao}
\usage{
rnp_metricas_regressao(observado, predito, digits = 4L)
}
\arguments{
\item{observado}{Vetor numerico observado.}

\item{predito}{Vetor numerico predito.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{rmse}, \code{mae}, \code{mape}, \code{r2}, \code{rmse_relativo}.
}
\description{
Metricas de regressao
}
\examples{
rnp_metricas_regressao(mtcars$mpg, predict(lm(mpg ~ wt, mtcars)))
}
\seealso{
Other avaliacao: 
\code{\link{rnp_acuracia_diagnostica}()},
\code{\link{rnp_brier}()},
\code{\link{rnp_calibracao}()},
\code{\link{rnp_comparar_roc}()},
\code{\link{rnp_curva_ganho}()},
\code{\link{rnp_curva_lift}()},
\code{\link{rnp_curva_precisao_revocacao}()},
\code{\link{rnp_ks_classificador}()},
\code{\link{rnp_metricas_classificacao}()}
}
\concept{avaliacao}
