% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_arvore}
\alias{rnp_ml_arvore}
\title{Especificacao de arvore de decisao}
\usage{
rnp_ml_arvore(
  modo = c("classificacao", "regressao"),
  custo_complexidade = 0.01,
  profundidade_max = 30,
  min_n = 2
)
}
\arguments{
\item{modo}{\code{"classificacao"} ou \code{"regressao"}.}

\item{custo_complexidade}{Parametro de poda (cp). Pode ser \code{tune()}.}

\item{profundidade_max}{Profundidade maxima.}

\item{min_n}{Minimo de observacoes por no.}
}
\value{
Especificacao \code{model_spec} (engine rpart).
}
\description{
Especificacao de arvore de decisao
}
\examples{
\donttest{ rnp_ml_arvore("classificacao") }
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
