% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_boosting}
\alias{rnp_ml_boosting}
\title{Especificacao de gradient boosting}
\usage{
rnp_ml_boosting(
  modo = c("classificacao", "regressao"),
  arvores = 500,
  profundidade = 6,
  taxa_aprendizado = 0.1
)
}
\arguments{
\item{modo}{\code{"classificacao"} ou \code{"regressao"}.}

\item{arvores}{Numero de arvores.}

\item{profundidade}{Profundidade das arvores.}

\item{taxa_aprendizado}{Taxa de aprendizado.}
}
\value{
Especificacao \code{model_spec} (engine xgboost).
}
\description{
Especificacao de gradient boosting
}
\examples{
\donttest{ rnp_ml_boosting("regressao") }
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
