% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_cv}
\alias{rnp_ml_cv}
\title{Reamostras de validacao cruzada}
\usage{
rnp_ml_cv(data, v = 10, repeticoes = 1, estrato = NULL, seed = 42L)
}
\arguments{
\item{data}{data.frame.}

\item{v}{Numero de folds.}

\item{repeticoes}{Numero de repeticoes.}

\item{estrato}{Nome (string) da variavel de estratificacao. Opcional.}

\item{seed}{Inteiro.}
}
\value{
Um objeto \code{rset} de reamostras.
}
\description{
Cria as particoes de validacao cruzada k-fold (\code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}).
}
\examples{
\donttest{
rnp_ml_cv(iris, v = 5, estrato = "Species")
}
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
