% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_floresta}
\alias{rnp_ml_floresta}
\title{Especificacao de floresta aleatoria}
\usage{
rnp_ml_floresta(
  modo = c("classificacao", "regressao"),
  arvores = 500,
  mtry = NULL,
  min_n = 5
)
}
\arguments{
\item{modo}{\code{"classificacao"} ou \code{"regressao"}.}

\item{arvores}{Numero de arvores.}

\item{mtry}{Variaveis sorteadas por divisao. \code{NULL} usa o padrao.}

\item{min_n}{Minimo de observacoes por no.}
}
\value{
Especificacao \code{model_spec} (engine ranger).
}
\description{
Especificacao de floresta aleatoria
}
\examples{
\donttest{ rnp_ml_floresta("classificacao") }
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
