% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_knn}
\alias{rnp_ml_knn}
\title{Especificacao de k-vizinhos}
\usage{
rnp_ml_knn(modo = c("classificacao", "regressao"), vizinhos = 5)
}
\arguments{
\item{modo}{\code{"classificacao"} ou \code{"regressao"}.}

\item{vizinhos}{Numero de vizinhos.}
}
\value{
Especificacao \code{model_spec} (engine kknn).
}
\description{
Especificacao de k-vizinhos
}
\examples{
\donttest{ rnp_ml_knn("classificacao") }
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
