% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_particao}
\alias{rnp_ml_particao}
\title{Particao treino/teste}
\usage{
rnp_ml_particao(data, prop = 0.75, estrato = NULL, seed = 42L)
}
\arguments{
\item{data}{data.frame.}

\item{prop}{Proporcao para treino.}

\item{estrato}{Nome (string) da variavel de estratificacao. Opcional.}

\item{seed}{Inteiro. Semente.}
}
\value{
Um objeto de particao \code{rsplit} (use \code{\link[=rnp_ml_ajustar]{rnp_ml_ajustar()}} ou
\code{rsample::training()}/\code{testing()}).
}
\description{
Divide os dados em conjuntos de treino e teste (\code{\link[rsample:initial_split]{rsample::initial_split()}}).
}
\examples{
\donttest{
rnp_ml_particao(iris, prop = 0.75, estrato = "Species")
}
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
