% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_prever}
\alias{rnp_ml_prever}
\title{Predicao com modelo ajustado}
\usage{
rnp_ml_prever(
  modelo,
  novos_dados,
  tipo = c("classe", "probabilidade", "numerico")
)
}
\arguments{
\item{modelo}{Workflow ajustado (de \code{\link[=rnp_ml_ajustar]{rnp_ml_ajustar()}}) ou parsnip fit.}

\item{novos_dados}{data.frame.}

\item{tipo}{\code{"classe"}, \code{"probabilidade"} ou \code{"numerico"}.}
}
\value{
tibble de predicoes.
}
\description{
Predicao com modelo ajustado
}
\examples{
\donttest{
sp <- rnp_ml_particao(iris, estrato = "Species")
fit <- rnp_ml_ajustar(rnp_ml_arvore("classificacao"), Species ~ ., sp)
rnp_ml_prever(fit$modelo, iris[1:5, ], tipo = "classe")
}
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_regularizada}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
