% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{rnp_ml_regularizada}
\alias{rnp_ml_regularizada}
\title{Especificacao de modelo linear regularizado (glmnet)}
\usage{
rnp_ml_regularizada(
  modo = c("classificacao", "regressao"),
  penalidade = 0.1,
  mistura = 1
)
}
\arguments{
\item{modo}{\code{"classificacao"} ou \code{"regressao"}.}

\item{penalidade}{Forca da penalizacao (lambda). Pode ser \code{tune()}.}

\item{mistura}{Mistura L1/L2 (alpha) em [0, 1].}
}
\value{
Especificacao \code{model_spec} (engine glmnet).
}
\description{
Lasso/ridge/elastic net via glmnet (\code{mistura = 1} lasso; \code{0} ridge).
}
\examples{
\donttest{ rnp_ml_regularizada("regressao", penalidade = 0.1) }
}
\seealso{
Other ml: 
\code{\link{rnp_ml_ajustar}()},
\code{\link{rnp_ml_arvore}()},
\code{\link{rnp_ml_boosting}()},
\code{\link{rnp_ml_comparar}()},
\code{\link{rnp_ml_cv}()},
\code{\link{rnp_ml_floresta}()},
\code{\link{rnp_ml_importancia}()},
\code{\link{rnp_ml_knn}()},
\code{\link{rnp_ml_particao}()},
\code{\link{rnp_ml_prever}()},
\code{\link{rnp_ml_receita}()},
\code{\link{rnp_ml_svm}()},
\code{\link{rnp_ml_tunagem}()}
}
\concept{ml}
