% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processos.R
\name{rnp_monte_carlo}
\alias{rnp_monte_carlo}
\title{Integracao por Monte Carlo}
\usage{
rnp_monte_carlo(
  integrando,
  limites = c(0, 1),
  n = 10000L,
  conf = 0.95,
  seed = 42L,
  digits = 4L
)
}
\arguments{
\item{integrando}{Funcao vetorizada a integrar.}

\item{limites}{Vetor \code{c(a, b)} com os limites de integracao.}

\item{n}{Inteiro. Numero de pontos de Monte Carlo.}

\item{conf}{Nivel de confianca do IC.}

\item{seed}{Inteiro. Semente.}

\item{digits}{Inteiro. Casas decimais.}
}
\value{
tibble com \code{estimativa}, \code{erro_padrao}, \code{ic_inf}, \code{ic_sup}, \code{n}.
}
\description{
Estima a integral de \code{integrando} em \verb{\[a, b\]} por amostragem uniforme,
com erro-padrao e intervalo de confianca.
}
\examples{
# integral de x^2 em [0, 1] = 1/3
rnp_monte_carlo(function(x) x^2, limites = c(0, 1), n = 1e5)
}
\seealso{
Other probabilidade: 
\code{\link{rnp_bayes}()},
\code{\link{rnp_cadeia_markov}()},
\code{\link{rnp_distribuicao_conjunta}()},
\code{\link{rnp_esperanca_condicional}()},
\code{\link{rnp_lei_grandes_numeros}()},
\code{\link{rnp_passeio_aleatorio}()},
\code{\link{rnp_processo_poisson}()},
\code{\link{rnp_simula_aceitacao_rejeicao}()},
\code{\link{rnp_simula_inversao}()},
\code{\link{rnp_tcl_simulacao}()}
}
\concept{probabilidade}
