% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobrevivencia.R
\name{rnp_nelson_aalen}
\alias{rnp_nelson_aalen}
\title{Risco acumulado de Nelson-Aalen}
\usage{
rnp_nelson_aalen(tempo, evento, digits = 4L)
}
\arguments{
\item{tempo}{Vetor de tempos.}

\item{evento}{Vetor 0/1.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{tempo}, \code{n_risco}, \code{n_evento}, \code{risco_acumulado}, \code{ep}.
}
\description{
Estima a funcao de risco acumulado H(t) pelo estimador de Nelson-Aalen.
}
\examples{
head(rnp_nelson_aalen(survival::lung$time, survival::lung$status == 2))
}
\seealso{
Other sobrevivencia: 
\code{\link{rnp_cox}()},
\code{\link{rnp_cox_diagnosticos}()},
\code{\link{rnp_cox_risco_relativo}()},
\code{\link{rnp_grafico_sobrevivencia}()},
\code{\link{rnp_kaplan_meier}()},
\code{\link{rnp_log_rank}()},
\code{\link{rnp_sobrevivencia_parametrica}()},
\code{\link{rnp_tabela_vida}()}
}
\concept{sobrevivencia}
