% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada-avancada.R
\name{rnp_normalidade_multivariada}
\alias{rnp_normalidade_multivariada}
\title{Teste de normalidade multivariada (Mardia)}
\usage{
rnp_normalidade_multivariada(X, digits = 4L)
}
\arguments{
\item{X}{Matriz/data.frame numerica.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{medida}, \code{estatistica}, \code{p_valor}.
}
\description{
Avalia assimetria e curtose multivariadas (teste de Mardia).
}
\examples{
rnp_normalidade_multivariada(iris[, 1:4])
}
\seealso{
Other multivariada: 
\code{\link{rnp_analise_fatorial}()},
\code{\link{rnp_biplot}()},
\code{\link{rnp_cluster_hierarquico}()},
\code{\link{rnp_correlacao_canonica}()},
\code{\link{rnp_correspondencia}()},
\code{\link{rnp_grafico_correlograma}()},
\code{\link{rnp_grafico_dendrograma}()},
\code{\link{rnp_grafico_dispersao_matriz}()},
\code{\link{rnp_hotelling}()},
\code{\link{rnp_kmedoids}()},
\code{\link{rnp_lda}()},
\code{\link{rnp_manova}()},
\code{\link{rnp_matriz_correlacao}()},
\code{\link{rnp_silhueta}()}
}
\concept{multivariada}
