% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessamento.R
\name{rnp_normaliza}
\alias{rnp_normaliza}
\title{Normaliza (min-max para [0, 1])}
\usage{
rnp_normaliza(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vetor numerico.}

\item{na.rm}{Logico.}
}
\value{
Vetor numerico em [0, 1].
}
\description{
Reescala para o intervalo [0, 1]: (x - min) / (max - min).
}
\examples{
rnp_normaliza(mtcars$hp)
}
\seealso{
Other preprocessamento: 
\code{\link{rnp_discretiza}()},
\code{\link{rnp_dummy}()},
\code{\link{rnp_imputa}()},
\code{\link{rnp_padroniza}()},
\code{\link{rnp_winsoriza}()}
}
\concept{preprocessamento}
