% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp-package.R
\name{rnp_options}
\alias{rnp_options}
\title{Opcoes do pacote rnp}
\usage{
rnp_options(...)
}
\arguments{
\item{...}{nome=valor}
}
\value{
Lista invisivel de opcoes anteriores.
}
\description{
Opcoes do pacote rnp
}
