% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessamento.R
\name{rnp_padroniza}
\alias{rnp_padroniza}
\title{Padroniza (z-score)}
\usage{
rnp_padroniza(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vetor numerico.}

\item{na.rm}{Logico.}
}
\value{
Vetor numerico padronizado.
}
\description{
Centra na media e escala pelo desvio-padrao: z = (x - media) / dp.
}
\examples{
rnp_padroniza(mtcars$mpg)
}
\seealso{
Other preprocessamento: 
\code{\link{rnp_discretiza}()},
\code{\link{rnp_dummy}()},
\code{\link{rnp_imputa}()},
\code{\link{rnp_normaliza}()},
\code{\link{rnp_winsoriza}()}
}
\concept{preprocessamento}
