% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_paleta_rnp}
\alias{rnp_paleta_rnp}
\title{Paletas de cores do projeto R NA PRATICA}
\usage{
rnp_paleta_rnp(nome = c("rnp_qual", "rnp_seq", "rnp_div"), n = NULL)
}
\arguments{
\item{nome}{String: \code{"rnp_qual"}, \code{"rnp_seq"}, \code{"rnp_div"}.}

\item{n}{Inteiro. Numero de cores.}
}
\value{
Vetor de cores (hex).
}
\description{
Paletas qualitativa, sequencial e divergente.
}
\examples{
rnp_paleta_rnp("rnp_qual", 5)
rnp_paleta_rnp("rnp_seq", 10)
}
