% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_permutacao}
\alias{rnp_permutacao}
\title{Numero de permutacoes}
\usage{
rnp_permutacao(n, elementos = NULL)
}
\arguments{
\item{n}{Inteiro. Total de elementos.}

\item{elementos}{Vetor opcional de frequencias (permutacao com repeticao).}
}
\value{
Escalar numerico.
}
\description{
Numero de permutacoes
}
\examples{
rnp_permutacao(5)
rnp_permutacao(7, elementos = c(2, 2, 3))
}
