% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilidade.R
\name{rnp_probabilidade_condicional}
\alias{rnp_probabilidade_condicional}
\title{Probabilidade condicional}
\usage{
rnp_probabilidade_condicional(pa_e_b, pb)
}
\arguments{
\item{pa_e_b}{Numerico em [0, 1]. Probabilidade da interseccao.}

\item{pb}{Numerico em (0, 1]. Probabilidade do evento condicionante.}
}
\value{
Escalar numerico.
}
\description{
Calcula P(A|B) = P(A & B) / P(B).
}
\examples{
rnp_probabilidade_condicional(0.12, 0.30)
}
