% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{rnp_quantis}
\alias{rnp_quantis}
\title{Quantis com metodo configuravel}
\usage{
rnp_quantis(
  x,
  probs = c(0.25, 0.5, 0.75),
  type = 7L,
  na.rm = TRUE,
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico.}

\item{probs}{Vetor de probabilidades em [0, 1].}

\item{type}{Inteiro 1-9. Metodo do quantil.}

\item{na.rm}{Logico.}

\item{digits}{Inteiro. Casas decimais.}
}
\value{
Vetor nomeado.
}
\description{
Wrapper de \code{\link[stats:quantile]{stats::quantile()}} expondo o argumento \code{type}
(Hyndman-Fan, 1 a 9).
}
\examples{
rnp_quantis(1:100, c(.1, .25, .5, .75, .9), type = 7)
}
