% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_read}
\alias{rnp_read}
\title{Leitura rapida de bases de dados delimitadas}
\usage{
rnp_read(
  base,
  sep = "|",
  dec = ".",
  encoding = "Latin-1",
  nrows = Inf,
  verbose = TRUE,
  select = NULL,
  ...
)
}
\arguments{
\item{base}{Local e nome do arquivo.}

\item{sep}{Separador de colunas.}

\item{dec}{Separador decimal.}

\item{encoding}{Codificacao do arquivo.}

\item{nrows}{Numero de linhas a ler (\code{Inf} = todas).}

\item{verbose}{Logico. Exibe mensagens de progresso/diagnostico.}

\item{select}{Vetor de colunas de interesse (default: todas).}

\item{...}{Argumentos extras para \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
Um \link[tibble:tibble]{tibble}.
}
\description{
Wrapper de \code{\link[readr:read_delim]{readr::read_delim()}} otimizado para os microdados do INEP
(delimitador padrao \code{"|"} e codificacao \code{Latin-1}).
}
\examples{
\dontrun{
rnp_read("dados.csv", sep = "|", encoding = "Latin-1")
}
}
