% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao-avancada.R
\name{rnp_regressao_ponderada}
\alias{rnp_regressao_ponderada}
\title{Regressao linear ponderada (WLS)}
\usage{
rnp_regressao_ponderada(formula, data, pesos, digits = 4L)
}
\arguments{
\item{formula}{Formula.}

\item{data}{data.frame.}

\item{pesos}{Vetor de pesos (positivos), comprimento = nrow(data).}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{termo}, \code{estimativa}, \code{erro_padrao}, \code{estatistica_t},
\code{p_valor}.
}
\description{
Minimos quadrados ponderados via backend C++.
}
\examples{
rnp_regressao_ponderada(mpg ~ wt, mtcars, pesos = 1 / mtcars$wt)
}
\seealso{
Other regressao: 
\code{\link{rnp_anova_modelos}()},
\code{\link{rnp_box_cox}()},
\code{\link{rnp_elastic_net}()},
\code{\link{rnp_grafico_residuos}()},
\code{\link{rnp_predicao}()},
\code{\link{rnp_regressao_lasso}()},
\code{\link{rnp_regressao_multinomial}()},
\code{\link{rnp_regressao_nao_linear}()},
\code{\link{rnp_regressao_poisson}()},
\code{\link{rnp_regressao_polinomial}()},
\code{\link{rnp_regressao_ridge}()},
\code{\link{rnp_regressao_robusta}()},
\code{\link{rnp_regressao_stepwise}()},
\code{\link{rnp_vif}()}
}
\concept{regressao}
