% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_sarima}
\alias{rnp_sarima}
\title{Ajuste de modelo SARIMA (ARIMA sazonal)}
\usage{
rnp_sarima(
  x,
  ordem = c(0, 0, 0),
  sazonal = c(0, 0, 0),
  periodo = 12,
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico ou \code{ts}.}

\item{ordem}{Vetor \code{c(p, d, q)} da parte nao-sazonal.}

\item{sazonal}{Vetor \code{c(P, D, Q)} da parte sazonal.}

\item{periodo}{Periodo sazonal (ex.: 12 para mensal).}

\item{digits}{Inteiro.}
}
\value{
Mesma estrutura de \code{\link[=rnp_arima]{rnp_arima()}}.
}
\description{
Ajuste de modelo SARIMA (ARIMA sazonal)
}
\examples{
rnp_sarima(AirPassengers, c(0, 1, 1), c(0, 1, 1), 12)$modelo
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
