% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_scheffe}
\alias{rnp_scheffe}
\title{Teste de Scheffe}
\usage{
rnp_scheffe(modelo, data = NULL, conf = 0.95, digits = 4L)
}
\arguments{
\item{modelo}{Objeto \code{aov} ou formula.}

\item{data}{data.frame (se modelo for formula).}

\item{conf}{Nivel de confianca.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Teste de Scheffe
}
