% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{rnp_skewness}
\alias{rnp_skewness}
\title{Coeficiente de assimetria (skewness)}
\usage{
rnp_skewness(x, method = c("fisher", "pearson", "bowley"), na.rm = TRUE)
}
\arguments{
\item{x}{Vetor numerico.}

\item{method}{String: \code{"fisher"}, \code{"pearson"} ou \code{"bowley"}.}

\item{na.rm}{Logico.}
}
\value{
Escalar numerico.
}
\description{
Calcula o coeficiente de assimetria. Metodos: \code{pearson},
\code{fisher} (bias-corrected, default), \code{bowley} (quartis).
}
\examples{
rnp_skewness(rnorm(1000))
rnp_skewness(1:100, method = "bowley")
}
