% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_summary_all}
\alias{rnp_summary_all}
\title{Sumario estatistico geral}
\usage{
rnp_summary_all(base)
}
\arguments{
\item{base}{data.frame de entrada}
}
\value{
data.frame com as estatisticas por variaveis.
}
\description{
Sumario estatistico geral
}
\details{
Calcula estatisticas descritivas para todos os vetores numpericos de uma base de dados.
Se tiver variaveis categoricas, a funcao os ignora.
}
\examples{
\dontrun{
require(rnp)
data(dm_ies)
tmp <- rnp_summary_all(dm_ies$VL_RECEITA_PROPRIA)
# Variaveis numericas
head(tmp$num)
# Variaveis categoricas
head(tmp$cat)
}
}
\author{
LOPES, J. E
}
