% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_summary_by}
\alias{rnp_summary_by}
\title{Estatisticas descritivas por grupo}
\usage{
rnp_summary_by(base, variavel, grupos, digits = 3)
}
\arguments{
\item{base}{data.frame com as variaveis de entrada}

\item{variavel}{o nome da variavel numerica entre aspas}

\item{grupos}{lista ou vetor de nomes das estatisticas de agrupamento na ordem em que deseja
obter os resultados.}

\item{digits}{total de digitos para arredondar os decimais.}
}
\value{
As estatisticas de saida desta funcao sao: total (N), soma, numero de missing (Nmis),
minimo, primeiro quartil (Q1), mediana (Q2), terceiro quartil (Q3), maximo, desvio pacrao (devpad)
e coeficiente de variacao (cv) em um data.frame agrupadas conforme as classes das variaveis de
agrupamento.
}
\description{
Calcula estatisticas descritivas por grupo. Ela recebe como entrada um data.frame
o nome da variavel numerica e um vetor ou lista de nomes das variaveis que serao utilizadas
como grupos. A funcao trabalha com apoio da funcao \code{plyr::\link[plyr]{ddply}} e aceita muitos
grupos.
}
\examples{
rnp_summary_by(base = mtcars, variavel = "wt", grupos = "gear")
rnp_summary_by(base = mtcars, variavel = "wt", grupos = c("gear","cyl"))
rnp_summary_by(base = mtcars, variavel = "wt",
               grupos = list("gear","cyl"), digits = 2)
}
\author{
LOPES, J. E.
}
