% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencias.R
\name{rnp_tabela_frequencia}
\alias{rnp_tabela_frequencia}
\title{Tabela de frequencias para variavel categorica}
\usage{
rnp_tabela_frequencia(x, ordenar = FALSE, digits = 4L)
}
\arguments{
\item{x}{Vetor categorico, fator ou caractere.}

\item{ordenar}{Logico. Ordena por frequencia decrescente.}

\item{digits}{Inteiro. Casas decimais.}
}
\value{
tibble com \code{categoria}, \code{fa}, \code{fr}, \code{fa_acumulada}, \code{fr_acumulada}.
}
\description{
Constroi tabela de frequencias absolutas e relativas (simples e
acumuladas) para um vetor categorico ou fator.
}
\examples{
rnp_tabela_frequencia(mtcars$cyl)
rnp_tabela_frequencia(letters[sample(3, 50, TRUE)], ordenar = TRUE)
}
\seealso{
Other descritiva: 
\code{\link{rnp_intervalo_classes}()},
\code{\link{rnp_medias}()},
\code{\link{rnp_momentos}()},
\code{\link{rnp_tabela_contingencia}()}
}
\concept{descritiva}
