% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amostragem.R
\name{rnp_tamanho_amostra_media}
\alias{rnp_tamanho_amostra_media}
\title{Tamanho de amostra para media}
\usage{
rnp_tamanho_amostra_media(sigma, E, conf = 0.95, N = Inf)
}
\arguments{
\item{sigma}{Desvio-padrao populacional (estimativa).}

\item{E}{Margem de erro desejada.}

\item{conf}{Nivel de confianca.}

\item{N}{Tamanho da populacao (default Inf = populacao infinita).}
}
\value{
tibble com \code{n_infinita}, \code{n_finita}, \code{E}, \code{conf}.
}
\description{
Calcula n necessario para estimar media populacional com margem de erro E.
}
\examples{
rnp_tamanho_amostra_media(sigma = 10, E = 2, conf = 0.95)
}
