% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amostragem.R
\name{rnp_tamanho_amostra_proporcao}
\alias{rnp_tamanho_amostra_proporcao}
\title{Tamanho de amostra para proporcao}
\usage{
rnp_tamanho_amostra_proporcao(p = 0.5, E, conf = 0.95, N = Inf)
}
\arguments{
\item{p}{Proporcao esperada (default 0.5 = maximo).}

\item{E}{Margem de erro.}

\item{conf}{Nivel de confianca.}

\item{N}{Tamanho da populacao (default Inf).}
}
\value{
tibble.
}
\description{
Tamanho de amostra para proporcao
}
\examples{
rnp_tamanho_amostra_proporcao(E = 0.05, conf = 0.95)
}
