% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rnp_tamanho_efeito}
\alias{rnp_tamanho_efeito}
\title{Tamanho de efeito}
\usage{
rnp_tamanho_efeito(x, y = NULL, paired = FALSE, digits = 4L)
}
\arguments{
\item{x, y}{Vetores numericos (para d, g, r) ou objeto \code{aov} (para eta2).}

\item{paired}{Logico. Amostras pareadas (apenas para d, g, r).}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{cohens_d}, \code{hedges_g}, \code{r}, \code{eta2}.
}
\description{
Calcula medidas de tamanho de efeito: Cohen's d, Hedges' g, r, eta-quadrado.
}
\examples{
rnp_tamanho_efeito(rnorm(30, 5), rnorm(30, 5.5))
}
