% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rnp_tema_rnp}
\alias{rnp_tema_rnp}
\title{Tema ggplot2 do projeto R NA PRATICA}
\usage{
rnp_tema_rnp(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Escalar. Tamanho base da fonte.}

\item{base_family}{String. Familia de fonte.}
}
\value{
Objeto \code{theme} do ggplot2.
}
\description{
Tema minimalista com cores e tipografia educacionais.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + rnp_tema_rnp()
}
