% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes-inferencia.R
\name{rnp_teste_grubbs}
\alias{rnp_teste_grubbs}
\title{Teste de Grubbs para outlier}
\usage{
rnp_teste_grubbs(x, digits = 4L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{valor_extremo}, \code{estatistica}, \code{p_valor}.
}
\description{
Detecta um unico outlier (o valor mais extremo) sob normalidade.
}
\examples{
rnp_teste_grubbs(c(rnorm(30), 8))
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
