% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_teste_kendall_w}
\alias{rnp_teste_kendall_w}
\title{Coeficiente de concordancia de Kendall (W)}
\usage{
rnp_teste_kendall_w(matriz, digits = 4L)
}
\arguments{
\item{matriz}{Matriz com linhas = objetos, colunas = avaliadores.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Coeficiente de concordancia de Kendall (W)
}
