% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_teste_mcnemar}
\alias{rnp_teste_mcnemar}
\title{Teste de McNemar (pareado binario)}
\usage{
rnp_teste_mcnemar(x, y, digits = 4L)
}
\arguments{
\item{x, y}{Vetores binarios (0/1, TRUE/FALSE, ou factor 2 niveis).}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Teste de McNemar (pareado binario)
}
