% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes-inferencia.R
\name{rnp_teste_normalidade}
\alias{rnp_teste_normalidade}
\title{Testes de normalidade (Shapiro, Jarque-Bera, Anderson-Darling)}
\usage{
rnp_teste_normalidade(
  x,
  metodo = c("shapiro", "jarque_bera", "anderson_darling"),
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico.}

\item{metodo}{String: \code{"shapiro"}, \code{"jarque_bera"} ou \code{"anderson_darling"}.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{estatistica}, \code{p_valor}, \code{metodo}.
}
\description{
Testes de normalidade (Shapiro, Jarque-Bera, Anderson-Darling)
}
\examples{
set.seed(1)
rnp_teste_normalidade(rnorm(200), "jarque_bera")
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_razao_veross}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
