% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verossimilhanca.R
\name{rnp_teste_razao_veross}
\alias{rnp_teste_razao_veross}
\title{Teste da razao de verossimilhancas (modelos aninhados)}
\usage{
rnp_teste_razao_veross(modelo_completo, modelo_reduzido, digits = 4L)
}
\arguments{
\item{modelo_completo, modelo_reduzido}{Objetos \code{lm}/\code{glm} aninhados.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{estatistica}, \code{gl}, \code{p_valor}.
}
\description{
Compara dois modelos aninhados (\code{lm}/\code{glm}) pela estatistica
2 (logL_completo - logL_reduzido) ~ qui-quadrado.
}
\examples{
m0 <- lm(mpg ~ wt, mtcars)
m1 <- lm(mpg ~ wt + hp, mtcars)
rnp_teste_razao_veross(m1, m0)
}
\seealso{
Other inferencia: 
\code{\link{rnp_bayes_conjugada}()},
\code{\link{rnp_bootstrap}()},
\code{\link{rnp_bootstrap_parametrico}()},
\code{\link{rnp_emv}()},
\code{\link{rnp_ic_bootstrap}()},
\code{\link{rnp_ic_verossimilhanca}()},
\code{\link{rnp_informacao_fisher}()},
\code{\link{rnp_jackknife}()},
\code{\link{rnp_log_verossimilhanca}()},
\code{\link{rnp_metodo_momentos}()},
\code{\link{rnp_poder_teste}()},
\code{\link{rnp_tamanho_amostra_teste}()},
\code{\link{rnp_teste_aderencia}()},
\code{\link{rnp_teste_binomial}()},
\code{\link{rnp_teste_grubbs}()},
\code{\link{rnp_teste_ks}()},
\code{\link{rnp_teste_normalidade}()},
\code{\link{rnp_teste_permutacao}()},
\code{\link{rnp_teste_proporcoes}()},
\code{\link{rnp_teste_qui_quadrado}()},
\code{\link{rnp_teste_runs}()},
\code{\link{rnp_teste_score}()},
\code{\link{rnp_teste_sinais}()},
\code{\link{rnp_teste_wald}()}
}
\concept{inferencia}
