% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencia.R
\name{rnp_teste_shapiro}
\alias{rnp_teste_shapiro}
\title{Teste de Shapiro-Wilk (normalidade)}
\usage{
rnp_teste_shapiro(x, digits = 4L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{estatistica}, \code{gl}, \code{p_valor}, \code{metodo}.
}
\description{
Teste de Shapiro-Wilk (normalidade)
}
\examples{
rnp_teste_shapiro(rnorm(50))
}
