% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencia.R
\name{rnp_teste_z_media}
\alias{rnp_teste_z_media}
\title{Teste Z para media (sigma conhecido)}
\usage{
rnp_teste_z_media(
  x,
  mu = 0,
  sigma,
  lado = c("bilateral", "direita", "esquerda"),
  conf = 0.95,
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico.}

\item{mu}{Media sob H0.}

\item{sigma}{Desvio-padrao populacional conhecido.}

\item{lado}{String: bilateral, direita ou esquerda.}

\item{conf}{Nivel de confianca.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Teste Z para media (sigma conhecido)
}
\examples{
rnp_teste_z_media(rnorm(100, mean = 10, sd = 2), mu = 10, sigma = 2)
}
