% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencia.R
\name{rnp_teste_z_proporcao}
\alias{rnp_teste_z_proporcao}
\title{Teste Z para proporcao (uma amostra)}
\usage{
rnp_teste_z_proporcao(
  sucesso,
  n,
  p0 = 0.5,
  lado = c("bilateral", "direita", "esquerda"),
  conf = 0.95,
  digits = 4L
)
}
\arguments{
\item{sucesso}{Numerico. Numero de sucessos.}

\item{n}{Inteiro. Tamanho da amostra.}

\item{p0}{Proporcao sob H0 (default 0.5).}

\item{lado}{String.}

\item{conf}{Nivel de confianca.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Teste Z para proporcao (uma amostra)
}
\examples{
rnp_teste_z_proporcao(55, 100, p0 = 0.5)
}
