% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_try_error}
\alias{rnp_try_error}
\title{Trata erros}
\usage{
rnp_try_error(code, silent = TRUE)
}
\arguments{
\item{code}{Qualquer codigo R passado}

\item{silent}{Rodar silenciosamente? TRUE ou FALSE}
}
\value{
Objeto de entrada
}
\description{
Trata qualquer erro em chamadas de funcoes. Se der erro, sai um objeto da classe 'try-error'
}
\author{
LOPES, J. E
}
