% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_ts_acf_pacf}
\alias{rnp_ts_acf_pacf}
\title{ACF e PACF}
\usage{
rnp_ts_acf_pacf(x, lag.max = NULL, plot = TRUE, digits = 4L)
}
\arguments{
\item{x}{Vetor numerico ou objeto \code{ts}.}

\item{lag.max}{Inteiro. Numero maximo de lags.}

\item{plot}{Logico. Gerar grafico.}

\item{digits}{Inteiro.}
}
\value{
lista:
\itemize{
\item \code{acf}: tibble com lag, acf.
\item \code{pacf}: tibble com lag, pacf.
\item \code{plot}: objeto ggplot (se plot = TRUE).
}
}
\description{
Calcula e plota ACF e PACF com IC 95\%.
}
\examples{
rnp_ts_acf_pacf(rnorm(100), lag.max = 20)
}
