% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_ts_adf}
\alias{rnp_ts_adf}
\title{Teste de Dickey-Fuller aumentado (ADF)}
\usage{
rnp_ts_adf(x, lag = NULL, tipo = c("constante", "tendencia"), digits = 4L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{lag}{Numero de defasagens aumentadas. Default automatico.}

\item{tipo}{String: \code{"constante"} (default) ou \code{"tendencia"}.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{estatistica}, \code{lag}, \code{valor_critico_5}, \code{p_valor_aprox},
\code{estacionaria}.
}
\description{
Testa a presenca de raiz unitaria (H0: serie nao-estacionaria). Ajusta a
regressao da serie diferenciada sobre o nivel defasado e sobre defasagens da
propria diferenca (forma "aumentada"), e compara a estatistica do termo de
nivel com valores criticos de Dickey-Fuller.
}
\examples{
rnp_ts_adf(as.numeric(lh))
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
