% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_ts_ccf}
\alias{rnp_ts_ccf}
\title{Funcao de correlacao cruzada (CCF)}
\usage{
rnp_ts_ccf(x, y, lag_max = NULL, digits = 4L)
}
\arguments{
\item{x, y}{Vetores numericos.}

\item{lag_max}{Numero maximo de defasagens.}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{tabela} (tibble) e \code{grafico} (\code{ggplot}).
}
\description{
Funcao de correlacao cruzada (CCF)
}
\examples{
rnp_ts_ccf(as.numeric(mdeaths), as.numeric(fdeaths))$tabela
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
