% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_ts_decomposicao}
\alias{rnp_ts_decomposicao}
\title{Decomposicao de serie temporal}
\usage{
rnp_ts_decomposicao(
  x,
  type = c("classica", "stl"),
  frequency = NULL,
  digits = 4L
)
}
\arguments{
\item{x}{Vetor numerico ou objeto \code{ts}.}

\item{type}{String: \code{"classica"} ou \code{"stl"}.}

\item{frequency}{Inteiro. Periodicidade (default \code{NULL} = auto).}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{serie} (objeto ajustado) e \code{componentes} (tibble).
}
\description{
Wrapper de \code{\link[stats:decompose]{stats::decompose()}} e \code{\link[stats:stl]{stats::stl()}}.
}
\examples{
x <- ts(rnorm(100) + rep(1:4, 25), frequency = 4)
rnp_ts_decomposicao(x)
}
\seealso{
Other series: 
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()}
}
\concept{series}
