% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_ts_diferenciacao}
\alias{rnp_ts_diferenciacao}
\title{Diferenciacao de serie}
\usage{
rnp_ts_diferenciacao(x, d = 1L, D = 0L, s = 12L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{d}{Inteiro. Ordem da diferenciacao regular.}

\item{D}{Inteiro. Ordem da diferenciacao sazonal.}

\item{s}{Inteiro. Periodo sazonal.}
}
\value{
Vetor numerico diferenciado.
}
\description{
Aplica diferenciacao regular (ordem \code{d}) e/ou sazonal (ordem \code{D}, periodo
\code{s}).
}
\examples{
rnp_ts_diferenciacao(as.numeric(AirPassengers), d = 1, D = 1, s = 12)
}
\seealso{
Other series: 
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()}
}
\concept{series}
