% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_ts_kpss}
\alias{rnp_ts_kpss}
\title{Teste KPSS de estacionariedade}
\usage{
rnp_ts_kpss(x, tipo = c("nivel", "tendencia"), digits = 4L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{tipo}{String: \code{"nivel"} (default) ou \code{"tendencia"}.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{estatistica}, \code{valor_critico_5}, \code{estacionaria}.
}
\description{
Testa H0 de \strong{estacionariedade} (oposto do ADF) em torno de um nivel
(ou tendencia), usando a soma acumulada dos residuos e variancia de longo
prazo de Newey-West.
}
\examples{
rnp_ts_kpss(as.numeric(lh))
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
