% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_temporais.R
\name{rnp_ts_periodograma}
\alias{rnp_ts_periodograma}
\title{Periodograma (analise espectral)}
\usage{
rnp_ts_periodograma(x, digits = 4L)
}
\arguments{
\item{x}{Vetor numerico.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{frequencia}, \code{periodo} e \code{densidade}.
}
\description{
Estima o periodograma (densidade espectral bruta) de uma serie.
}
\examples{
rnp_ts_periodograma(as.numeric(AirPassengers))
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
