% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_ts_previsao}
\alias{rnp_ts_previsao}
\title{Previsao a partir de modelo ARIMA/SARIMA}
\usage{
rnp_ts_previsao(modelo, h = 10, conf = 0.95, digits = 4L)
}
\arguments{
\item{modelo}{Saida de \code{\link[=rnp_arima]{rnp_arima()}}/\code{\link[=rnp_sarima]{rnp_sarima()}} (ou objeto \code{Arima}).}

\item{h}{Inteiro. Horizonte de previsao.}

\item{conf}{Nivel de confianca dos intervalos.}

\item{digits}{Inteiro.}
}
\value{
Uma lista com \code{previsao} (tibble) e \code{grafico} (\code{ggplot}).
}
\description{
Previsao a partir de modelo ARIMA/SARIMA
}
\examples{
m <- rnp_arima(lh, c(1, 0, 0))
rnp_ts_previsao(m, h = 6)$previsao
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_residuos}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
