% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-arima.R
\name{rnp_ts_residuos}
\alias{rnp_ts_residuos}
\title{Diagnostico de residuos de modelo de series}
\usage{
rnp_ts_residuos(modelo, lag = 10, digits = 4L)
}
\arguments{
\item{modelo}{Saida de \code{\link[=rnp_arima]{rnp_arima()}}/\code{\link[=rnp_sarima]{rnp_sarima()}} (ou objeto \code{Arima}).}

\item{lag}{Defasagens para o Ljung-Box.}

\item{digits}{Inteiro.}
}
\value{
tibble com os testes de diagnostico.
}
\description{
Ljung-Box e normalidade dos residuos de um ajuste ARIMA/SARIMA.
}
\examples{
rnp_ts_residuos(rnp_arima(lh, c(1, 0, 0)))
}
\seealso{
Other series: 
\code{\link{rnp_arima}()},
\code{\link{rnp_auto_arima}()},
\code{\link{rnp_grafico_acf}()},
\code{\link{rnp_grafico_serie}()},
\code{\link{rnp_media_movel}()},
\code{\link{rnp_sarima}()},
\code{\link{rnp_suavizacao_exponencial}()},
\code{\link{rnp_ts_acf}()},
\code{\link{rnp_ts_adf}()},
\code{\link{rnp_ts_ccf}()},
\code{\link{rnp_ts_decomposicao}()},
\code{\link{rnp_ts_diferenciacao}()},
\code{\link{rnp_ts_garch}()},
\code{\link{rnp_ts_holt_winters}()},
\code{\link{rnp_ts_kpss}()},
\code{\link{rnp_ts_ljung_box}()},
\code{\link{rnp_ts_pacf}()},
\code{\link{rnp_ts_periodograma}()},
\code{\link{rnp_ts_previsao}()},
\code{\link{rnp_ts_var}()}
}
\concept{series}
