% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_tukey_hsd}
\alias{rnp_tukey_hsd}
\title{Teste de Tukey HSD}
\usage{
rnp_tukey_hsd(modelo, data = NULL, conf = 0.95, digits = 4L)
}
\arguments{
\item{modelo}{Objeto \code{aov} ou formula.}

\item{data}{data.frame (se modelo for formula).}

\item{conf}{Nivel de confianca.}

\item{digits}{Inteiro.}
}
\value{
tibble com comparacoes.
}
\description{
Teste de Tukey HSD
}
\examples{
fit <- aov(mpg ~ factor(cyl), mtcars)
rnp_tukey_hsd(fit)
}
