% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao-avancada.R
\name{rnp_vif}
\alias{rnp_vif}
\title{Fator de inflacao de variancia (VIF)}
\usage{
rnp_vif(modelo, digits = 4L)
}
\arguments{
\item{modelo}{Objeto \code{lm}.}

\item{digits}{Inteiro.}
}
\value{
tibble com \code{termo}, \code{vif} e \code{interpretacao}.
}
\description{
Mede a multicolinearidade: VIF_j = 1 / (1 - R_j^2), onde R_j^2 e o R-quadrado
da regressao do preditor j sobre os demais.
}
\examples{
rnp_vif(lm(mpg ~ wt + hp + disp, mtcars))
}
\seealso{
Other regressao: 
\code{\link{rnp_anova_modelos}()},
\code{\link{rnp_box_cox}()},
\code{\link{rnp_elastic_net}()},
\code{\link{rnp_grafico_residuos}()},
\code{\link{rnp_predicao}()},
\code{\link{rnp_regressao_lasso}()},
\code{\link{rnp_regressao_multinomial}()},
\code{\link{rnp_regressao_nao_linear}()},
\code{\link{rnp_regressao_poisson}()},
\code{\link{rnp_regressao_polinomial}()},
\code{\link{rnp_regressao_ponderada}()},
\code{\link{rnp_regressao_ridge}()},
\code{\link{rnp_regressao_robusta}()},
\code{\link{rnp_regressao_stepwise}()}
}
\concept{regressao}
