% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rnp_wilcoxon}
\alias{rnp_wilcoxon}
\title{Teste de Wilcoxon para amostras pareadas}
\usage{
rnp_wilcoxon(x, y, lado = c("bilateral", "direita", "esquerda"), digits = 4L)
}
\arguments{
\item{x, y}{Vetores numericos pareados.}

\item{lado}{String.}

\item{digits}{Inteiro.}
}
\value{
tibble.
}
\description{
Atalho para Wilcoxon signed-rank.
}
\examples{
rnp_wilcoxon(rnorm(20, 5), rnorm(20, 5.2))
}
