% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\docType{data}
\name{dm_curso}
\alias{dm_curso}
\title{Dados do censo ensino superior INEP 2017 (Cursos)}
\description{
Este conjunto de dados faz parte da base de dados do INEP para o censo
brasileiro do ensino superior. Nesta parte temos os dados dos cursos
mapeados pelo censo naquele ano.
Trata-se de uma amostras das variáveis, porém o total de observações
está completo.
Para obter dados de outros anos consulte o link do site do INEP ou
a função \code{\link{rnp_get_inep_censo}}
}
\examples{
require(rnp)
str(dm_curso)
}
\references{
\url{http://inep.gov.br/microdados}
}
\author{
INEP
}
\keyword{dm_curso}
