% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_2freq}
\alias{rnp_2freq}
\title{Tabela de frequencias dupla entrada}
\usage{
rnp_2freq(x, y, digits = 4, percents = FALSE)
}
\arguments{
\item{x}{vetor caractere/fator de entrada}

\item{y}{vetor caractere/fator de entrada}

\item{digits}{total de digitos decimais na saida}

\item{percents}{se TRUE, retorna tambem os percentuais de linha e de colunas.}
}
\value{
data.frame
}
\description{
Tabela de frequencias dupla entrada
}
\details{
faz tabela de dupla entrada com base em dois vetores de entrada.
}
\author{
LOPES, J. E
}
