% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_associacao}
\alias{rnp_associacao}
\title{Estatística de associação}
\usage{
rnp_associacao(x, y, ...)
}
\arguments{
\item{x}{variável um}

\item{y}{variável dois}

\item{...}{argumentos passados para a função \code{\link{chisq.test}}}
}
\description{
Recebe dois vetores categóricos de tamanhos iguais e calcula as estatísticas
Qui-quadrado, V de Cramér e o coeficiante de contingência para a tabela de contingência
das duas variáveis criada pela função \code{\link{table}}.
}
\author{
LOPES, J. E.
}
