% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_associacao}
\alias{rnp_associacao}
\title{Estatistica de associacao}
\usage{
rnp_associacao(x, y, ...)
}
\arguments{
\item{x}{variavel um}

\item{y}{variavel dois}

\item{...}{argumentos passados para a funcao \code{\link{chisq.test}}}
}
\description{
Recebe dois vetores categoricos de tamanhos iguais e calcula as estatisticas
Qui-quadrado, V de Cramer e o coeficiante de contingencia para a tabela de contingencia
das duas variaveis criada pela funcao \code{\link{table}}.
}
\author{
LOPES, J. E.
}
