% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_atributos}
\alias{rnp_atributos}
\title{Extrai atributos de um objeto}
\usage{
rnp_atributos(obj, ests = FALSE)
}
\arguments{
\item{obj}{qualquer objeto em R}

\item{ests}{Se TRUE, retorna estatisticas do vetor ou das variaveis}
}
\value{
data.frame com lista de atributos
}
\description{
Extrai atributos de um objeto
}
\author{
LOPES, J. E
}
