% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_correlacao}
\alias{rnp_correlacao}
\title{Calculo de correlacao}
\usage{
rnp_correlacao(base, digits = 4)
}
\arguments{
\item{base}{data.frame, tibla, data.table, etc.}

\item{digits}{numeros de digitos para arrendondar o valor da correlacao.}

\item{...}{argumentos passados para a funcao \code{\link{chisq.test}}}
}
\value{
Um data.frame com cinco colunas onde as duas primeiras sao as combinacoes
aos pares de cada uma das variaveis e as tres ultimas sao as correlacoes de
Pearson, Spearman e Kendal respectivamente.
}
\description{
Esta funcao recebe como entrada uma base de dados com pelo menos duas
variaveis numericas e deternina os coeficientes de correlacao de Pearson, Spearman
e Kendal.
}
\details{
E possivel que existam variaveis nao numerias em sua base de dados. Neste caso,
elas serao eliminadas. Para mais detalhes veja \code{\link{cor}}.
}
\examples{
rnp_correlacao(mtcars)
}
\author{
LOPES, J. E.
}
