% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_freq}
\alias{rnp_freq}
\title{Tabela de frequências}
\usage{
rnp_freq(x, sortd = FALSE, digits = 4)
}
\arguments{
\item{x}{vetor numérico ou caractere de entrada}

\item{sortd}{TRUE para ordenar}
}
\value{
data.frame
}
\description{
Tabela de frequências
}
\details{
Calcula tabela de freência simples. Quando o vetor passado for caractere ou fator, a função trabalha com as frequências
e quando é passado um vetor numérico, ela utiliza função \code{\link{cut}} com apoio de  \code{\link{quantile}} para categorizar
os dados utilizando os quartis.
}
\author{
LOPES, J. E
}
