% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_get_classes_inep}
\alias{rnp_get_classes_inep}
\title{Obtem as classes das variáveis censo INEP}
\usage{
rnp_get_classes_inep(caminho, aba = 1, pula_linha = 1,
  retorna_lista = FALSE)
}
\arguments{
\item{caminho}{caminho do arquivo Excel do dicionário de dados do INEP}

\item{aba}{aba da planilha correspondente aos dados que deseja}

\item{pula_linha}{quantidade de linhas que deseja pular}

\item{retorna_lista}{TRUE se quer obter uma lista de data.frames, sendo
um para cada variável ou base já agregada.}
}
\description{
Obtem as classes das variáveis do diconário de dados do censo INEP.
}
\details{
Exige que passe o caminho para o dicionário de dados em excel e o nome da
aba correspondente ou o número da mesma, ex. 1 = primeira aba e assim por diante.
A função faz a leitura dos dados utilizando readxl::read_excel() e retorna
um data.frame com 4 aolunas contendo ordem, nome, categoria e descrição
para todas as variáveis da base informada.
OBS.: Testado apenas nos dados do censo de 2017.
}
\examples{
\dontrun{
nn <- c("DM_CURSO","DM_IES","DM_LOCAL_OFERTA","DM_DOCENTE")
L <- plyr::llply(nn[1], function(base){
  classes <- rnp_get_classes_inep(caminho = "Dicionário_de_Dados.xlsx",
                                  aba = base, retorna_lista = FALSE)
  base <- rnp_read(base = paste0("Dados/CSV/", base, ".CSV"),
                   sep = "|",
                   dec = ".",
                   header = TRUE,
                   encoding = "Latin-1",
                   verbose = FALSE,
                   showProgress = FALSE)
  oo <- rnp_aplica_classes(base = base, classes = classes)
  return(oo)
}, .progress = "text")
names(L) <- nn
}
}
\author{
LOPES, J. E.
}
