% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_get_inep_censo}
\alias{rnp_get_inep_censo}
\title{Download dados do INEP}
\usage{
rnp_get_inep_censo(ano = 2017, url = NULL, salvar = NULL)
}
\arguments{
\item{ano}{ano formato numerico, ex. 2010, 2016}

\item{url}{url, vetor ou lista de urls completas para baixar os dados.}

\item{salvar}{caminho onde deseja salvar os dados baixados. Se NULL, a funcao
baixa na pasta onde a sessao ativa do R foi iniciada. Execute \code{\link{getwd}} para saber o local.}
}
\value{
arquivo .zip ou da extencao da url passada.
}
\description{
Recebe uma data ou uma url do portal de microdados do INEP e
baixa os dados no local em que a sessao do R foi carregada.
}
\details{
Quando passado apenas uma data a funcao baixa os dados do INEP
da data correspondente para o censo de educacao superior, desde que esta
data seja entre 1995 e 2017. Caso a data esteja fora deste intervalo, a funcao
baixara os dados do censo de 2017.
Se uma url com final .zip for passada, a funcao ignora a data, caso tenha
sido informada e baixa o arquivo da url informada.
}
\examples{
\dontrun{
require(rnp)
require(dplyr)
url_in = 'http://download.inep.gov.br/microdados/micro_censo_edu_superior1995.zip'
rnp_get_inep_censo(url = url_in)
rnp_get_inep_censo()
}
}
\author{
LOPES, J. E.
}
