% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_read}
\alias{rnp_read}
\title{Leitura rapida de bases de dados}
\usage{
rnp_read(base, sep = "|", dec = ".", encoding = "Latin-1",
  nrows = Inf, verbose = TRUE, showProgress = TRUE, select = NULL,
  ...)
}
\arguments{
\item{base}{Local e nome com a extensao da base de dados de entrada}

\item{sep}{Separador de colunas}

\item{dec}{Separador decimal}

\item{encoding}{Codificacao dos dados, padrao e Latin-1. UTF-8 tambem e muito util}

\item{nrows}{Numero de linhas a extrair. Se Inf, tras todas}

\item{verbose}{Se TRUE, exibe logs do que esta aocontecendo em tempo de extracao}

\item{showProgress}{Se TRUE, exibe o andamento do processo de leitura}

\item{select}{Uma lista de variaveis de interesse pode ser passada para simplificar a saida.}

\item{...}{Passagem de argumentos ectras para a funcao \code{\link{fread}}}
}
\value{
tibla com os dados importados
}
\description{
Leitura rapida de bases de dados
}
\author{
LOPES, J. E
}
