% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_read}
\alias{rnp_read}
\title{Leitura rápida de bases de dados}
\usage{
rnp_read(base, sep = "|", dec = ".", encoding = "Latin-1",
  nrows = Inf, verbose = TRUE, showProgress = TRUE, ...)
}
\arguments{
\item{base}{Local e nome com a extensão da base de dados de entrada}

\item{sep}{Separador de colunas}

\item{dec}{Separador decimal}

\item{encoding}{Codificação dos dados, padrão é Latin-1. UTF-8 também é muito útil}

\item{nrows}{Número de linhas a extrair. Se Inf, trás todas}

\item{verbose}{Se TRUE, exibe logs do que está aocontecendo em tempo de extração}

\item{showProgress}{Se TRUE, exibe o andamento do processo de leitura}

\item{...}{Passagem de argumentos ectras para a função \code{\link{fread}}}
}
\value{
tibla com os dados importados
}
\description{
Leitura rápida de bases de dados
}
\author{
LOPES, J. E
}
