% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{rnp_summary}
\alias{rnp_summary}
\title{Sumario estatistico}
\usage{
rnp_summary(x, digits = 4)
}
\arguments{
\item{x}{vetor numerico de entrada}

\item{digits}{total de digitos decimais na saida}
}
\value{
Vetor nomeado
}
\description{
Sumario estatistico
}
\details{
Calcula estatisticas descritivas de um vetor numerico informado
}
\author{
LOPES, J. E
}
