% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_summary}
\alias{rnp_summary}
\title{Sumário estatístico}
\usage{
rnp_summary(x, digits = 4)
}
\arguments{
\item{x}{vetor numérico de entrada}

\item{digits}{total de digitos decimais na saida}
}
\value{
Vetor nomeado
}
\description{
Sumário estatístico
}
\details{
Calcula estatísticas descritivas de um vetor numérico informado
}
\author{
LOPES, J. E
}
