% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_summary_all}
\alias{rnp_summary_all}
\title{Sumário estatístico geral}
\usage{
rnp_summary_all(base)
}
\arguments{
\item{base}{data.frame de entrada}
}
\value{
data.frame com as estatísticas por variáveis.
}
\description{
Sumário estatístico geral
}
\details{
Calcula estatísticas descritivas para todos os vetores numpericos de uma base de dados.
Se tiver variáveis categóricas, a função os ignora.
}
\author{
LOPES, J. E
}
