% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_summary_all}
\alias{rnp_summary_all}
\title{Sumario estatistico geral}
\usage{
rnp_summary_all(base)
}
\arguments{
\item{base}{data.frame de entrada}
}
\value{
data.frame com as estatisticas por variaveis.
}
\description{
Sumario estatistico geral
}
\details{
Calcula estatisticas descritivas para todos os vetores numpericos de uma base de dados.
Se tiver variaveis categoricas, a funcao os ignora.
}
\author{
LOPES, J. E
}
