% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_FUNCOES.R
\name{rnp_summary_by}
\alias{rnp_summary_by}
\title{Estatísticas descritivas por grupo}
\usage{
rnp_summary_by(base, variavel, grupos, digits = 3)
}
\arguments{
\item{base}{data.frame com as variáveis de entrada}

\item{grupos}{lista ou vetor de nomes das estatísticas de agrupamento na ordem em que deseja
obter os resultados}

\item{variável}{o nome da variável numérica entre aspas}
}
\value{
As estatísticas de saída desta função são: total (N), soma, número de missing (Nmis),
minimo, primeiro quartil (Q1), mediana (Q2), terceiro quartil (Q3), máximo, desvio pacrão (devpad)
e coeficiente de variação (cv) em um data.frame agrupadas conforme as classes das variáveis de
agrupamento.
}
\description{
Calcula estatísticas descritivas por grupo. Ela recebe como entrada um data.frame
o nome da variável numérica e um vetor ou lista de nomes das variáveis que serão utilizadas
como grupos. A função trabalha com apoio da função \link{\code{plyr::ddply}} e aceita muitos
grupos.
}
\examples{
rnp_summary_by(base = mtcars, variavel = "wt", grupos = "gear")
rnp_summary_by(base = mtcars, variavel = "wt", grupos = c("gear","cyl"))
rnp_summary_by(base = mtcars, variavel = "wt",
               grupos = list("gear","cyl"), digits = 2)
}
\author{
LOPES, J. E.
}
